% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\docType{data}
\name{activities}
\alias{activities}
\title{R6 class activities creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
activities
}
\description{
Create R6 reference object class activities
}
\keyword{datasets}
