% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\docType{data}
\name{activity}
\alias{activity}
\title{R6 class activity creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
activity
}
\description{
Create R6 reference object class activity
}
\keyword{datasets}
