% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{activity_number}
\alias{activity_number}
\alias{check_activity_number}
\title{Attribut "activity_number" verification}
\usage{
check_activity_number(activity_number)
}
\arguments{
\item{activity_number}{(integer) Activity number}
}
\description{
Check if the item "activity_number" have one unique class identified as integer
}
