% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_ci_calculator.R
\name{catch_ci_calculator}
\alias{catch_ci_calculator}
\title{Compute final confidence interval}
\usage{
catch_ci_calculator(fit_data, boot_data)
}
\arguments{
\item{to}{The observed value of statistic applied to data}

\item{t}{A matrix with sum(R) rows each of which is a bootstrap replicate of the result of calling statistic.}
}
\value{

}
\description{
Compute final confidence interval
}
