% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{catch_weight}
\alias{catch_weight}
\alias{check_catch_weight}
\title{Attribut "catch_weight" verification}
\usage{
check_catch_weight(catch_weight)
}
\arguments{
\item{catch_weight}{(integer) Catch weight in tonne}
}
\description{
Check if the item "catch_weight" have one unique class identified as integer
}
