% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_activity_date}
\alias{check_activity_date}
\title{Attribute "activity_date" verification}
\usage{
check_activity_date(activity_date, landing_date = NULL)
}
\arguments{
\item{activity_date}{Object of class {\link[base]{character}} expected. Activity date in format year month day.}

\item{landing_date}{Object of class {\link[base]{character}} expected. Landing date in format year month day hour minute second UTC. By default NULL.}
}
\description{
Check if the item "activity_date" is in date format year month day, if the value is inferior to the actual date and if it is inferior to the landing_date (if there are provided).
}
