% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_activity_id}
\alias{check_activity_id}
\title{Attribut "activity_id" verification}
\usage{
check_activity_id(activity_id)
}
\arguments{
\item{activity_id}{(character) Activity identification}
}
\description{
Check if the item "activity_id" have one unique class identified as character
}
