% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_activity_latitude}
\alias{check_activity_latitude}
\title{Attribute "activity_latitude" verification}
\usage{
check_activity_latitude(activity_latitude)
}
\arguments{
\item{activity_latitude}{Object of class {\link[base]{numeric}} expected. Latitude, in decimal degree, of the activity.}
}
\description{
Check if the item "activity_latitude" have one unique class identified as numeric.
}
