% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_activity_longitude}
\alias{check_activity_longitude}
\title{Attribute "activity_longitude" verification}
\usage{
check_activity_longitude(activity_longitude)
}
\arguments{
\item{activity_longitude}{Object of class {\link[base]{numeric}} expected. Longitude, in decimal degree, of the activity.}
}
\description{
Check if the item "activity_longitude" have one unique class identified as numeric.
}
