% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_activity_name}
\alias{check_activity_name}
\title{Attribute "activity_name" verification}
\usage{
check_activity_name(activity_name)
}
\arguments{
\item{activity_name}{Object of class {\link[base]{character}} expected. Activity identification.}
}
\description{
Check if the item "check_activity_name" have one unique class identified as character.
}
