% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_catch_weight}
\alias{check_catch_weight}
\title{Attribute "catch_weight" verification}
\usage{
check_catch_weight(catch_weight)
}
\arguments{
\item{catch_weight}{Object of class {\link[base]{numeric}} expected. Catch weight in tonnes.}
}
\description{
Check if the item "catch_weight" have one unique class identified as numeric and if values are superior to zero.
}
