% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_departure_date}
\alias{check_departure_date}
\title{Attribute "departure_date" verification}
\usage{
check_departure_date(departure_date, landing_date = NULL)
}
\arguments{
\item{departure_date}{Object of class {\link[base]{character}} expected. Departure date in format year month day hour minute second UTC.}

\item{landing_date}{Object of class {\link[base]{character}} expected. Landing date in format year month day hour minute second UTC. By default NULL.}
}
\description{
Check if the item "departure_date" is in date format year month day hour minute second UTC, if the value is inferior to the actual date and if it is inferior to the landing_date (if there are provided).
}
