% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_departure_harbour_id}
\alias{check_departure_harbour_id}
\title{Attribut "departure_harbour_id" verification}
\usage{
check_departure_harbour_id(departure_harbour_id)
}
\arguments{
\item{departure_harbour_id}{(character) Departure harbour identification}
}
\description{
Check if the item "departure_harbour_id" have one unique class identified as character
}
