% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_elementarycatch_id}
\alias{check_elementarycatch_id}
\title{Attribute "elementarycatch_id" verification}
\usage{
check_elementarycatch_id(elementarycatch_id)
}
\arguments{
\item{elementarycatch_id}{Object of class {\link[base]{character}} expected. Elementary catch identification.}
}
\description{
Check if the item "elementarycatch_id" have one unique class identified as character.
}
