% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_elementarysampleraw_id}
\alias{check_elementarysampleraw_id}
\title{Attribute "elementarysampleraw_id" verification}
\usage{
check_elementarysampleraw_id(elementarysampleraw_id)
}
\arguments{
\item{elementarysampleraw_id}{Object of class {\link[base]{character}} expected. Elementary sample raw identification.}
}
\description{
Check if the item "elementarysampleraw_id" have one unique class identified as character.
}
