% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fish_hold_empty}
\alias{check_fish_hold_empty}
\title{Attribut "fish_hold_empty" verification}
\usage{
check_fish_hold_empty(fish_hold_empty)
}
\arguments{
\item{fish_hold_empty}{(integer) Informe if the fish hold empty at the end of the trip, 1 for yes and 0 for not}
}
\description{
Check if the item "fish_hold_empty" is class integer and if all values are 1 or 0
}
