% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fishing_time}
\alias{check_fishing_time}
\title{Attribut "fishing_time" verification}
\usage{
check_fishing_time(fishing_time)
}
\arguments{
\item{fishing_time}{(numeric or integer) Fishing time value}
}
\description{
Check if the item "fishing_time" is class integer or numeric and if all values are postives
}
