% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_landing_harbour_id}
\alias{check_landing_harbour_id}
\title{Attribut "landing_harbour_id" verification}
\usage{
check_landing_harbour_id(landing_harbour_id)
}
\arguments{
\item{landing_harbour_id}{(character) Landing harbour identification}
}
\description{
Check if the item "landing_harbour_id" have one unique class identified as character
}
