% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_length_type}
\alias{check_length_type}
\title{Attribut "length_type" verification}
\usage{
check_length_type(length_type)
}
\arguments{
\item{length_type}{(integer) Length type identification, 1 for LD1 and 2 for LF.}
}
\description{
Check if the item "length_type" is class integer and if all values are 1 or 2.
}
