% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_logbook_category}
\alias{check_logbook_category}
\title{Attribut "logbook_category" verification}
\usage{
check_logbook_category(logbook_category)
}
\arguments{
\item{logbook_category}{(integer) Logbook weight category}
}
\description{
Check if the item "logbook_category" have one unique class identified as integer and if values are between 1 and 13
}
