% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ocean}
\alias{check_ocean}
\title{Attribute "ocean" verification}
\usage{
check_ocean(ocean)
}
\arguments{
\item{ocean}{Object of class {\link[base]{integer}} expected. Ocean identification.}
}
\description{
Check if the item "ocean" is class integer and if all values are 1, 2, 3, 4 or 5.
}
