% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sample_number_measured_extrapolated_lf}
\alias{check_sample_number_measured_extrapolated_lf}
\title{Attribut "sample_number_measured_extrapolated_lf" verification}
\usage{
check_sample_number_measured_extrapolated_lf(
  sample_number_measured_extrapolated_lf
)
}
\arguments{
\item{sample_number_measured_extrapolated_lf}{(numeric) Sample number of measured individus extrapolated to all counted individus}
}
\description{
Check if the item "sample_number_measured_extrapolated_lf" have one unique class identified as numeric and if all values are positive and different of zero
}
