% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sample_quality}
\alias{check_sample_quality}
\title{Attribut "sample_quality" verification}
\usage{
check_sample_quality(sample_quality)
}
\arguments{
\item{sample_quality}{(integer) Sample quality identification}
}
\description{
Check if the item "sample_quality" have one unique class identified as integer and if all values are 1, 2, 3, 4 or 9
}
