% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sample_total_count}
\alias{check_sample_total_count}
\title{Attribute "sample_total_count" verification}
\usage{
check_sample_total_count(sample_total_count)
}
\arguments{
\item{sample_total_count}{Object of class {\link[base]{integer}} expected. Sample number of total individuals counted.}
}
\description{
Check if the item "sample_total_count" have one unique class identified as integer and if all values are positive and different of zero.
}
