% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_school_type}
\alias{check_school_type}
\title{Attribut "school_type" verification}
\usage{
check_school_type(school_type)
}
\arguments{
\item{school_type}{(integer) School type identification}
}
\description{
Check if the item "school_type" is class integer and if all values are 1, 2, or 3
}
