% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sub_sample_id_total_count}
\alias{check_sub_sample_id_total_count}
\title{Attribute "sub_sample_id_total_count" verification}
\usage{
check_sub_sample_id_total_count(sub_sample_id_total_count)
}
\arguments{
\item{sub_sample_id_total_count}{Object of class {\link[base]{character}} expected. Sub sample identification bis in relation with the fish total count.}
}
\description{
Check if the item "sub_sample_id_total_count" have one unique class identified as character.
}
