% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_time_at_sea}
\alias{check_time_at_sea}
\title{Attribut "time_at_sea" verification}
\usage{
check_time_at_sea(time_at_sea)
}
\arguments{
\item{time_at_sea}{(numeric or integer) Time at sea value}
}
\description{
Check if the item "time_at_sea" is class integer or numeric and if all values are postives
}
