% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_trip_id}
\alias{check_trip_id}
\title{Attribute "trip_id" verification}
\usage{
check_trip_id(trip_id)
}
\arguments{
\item{trip_id}{Object of class {\link[base]{character}} expected. Trip identification.}
}
\description{
Check if the item "trip_id" have one unique class identified as character.
}
