% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vessel_id}
\alias{check_vessel_id}
\title{Attribute "vessel_id" verification}
\usage{
check_vessel_id(vessel_id)
}
\arguments{
\item{vessel_id}{Object of class {\link[base]{integer}} expected. Vessel identification.}
}
\description{
Check if the item "vessel_id" have one unique class identified as integer.
}
