% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_weighted_weight}
\alias{check_weighted_weight}
\title{Attribut "weighted_weight" verification}
\usage{
check_weighted_weight(weighted_weight)
}
\arguments{
\item{weighted_weight}{(numeric) Set weight weighted by all set in the well(s)}
}
\description{
Check if the item "weighted_weight" have one unique class identified as numeric and if value are >= 0
}
