% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_well_global_weigth}
\alias{check_well_global_weigth}
\title{Attribute "well_global_weigth" verification}
\usage{
check_well_global_weigth(well_global_weigth)
}
\arguments{
\item{well_global_weigth}{Object of class {\link[base]{integer}} expected. Catch weight of individus (less and more 10 tonnes categories, by well, in tonne, all species considerated).}
}
\description{
Check if the item "well_global_weigth" have one unique class identified as integer and if value are >= 0.
}
