% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_well_id}
\alias{check_well_id}
\title{Attribute "well_id" verification}
\usage{
check_well_id(well_id)
}
\arguments{
\item{well_id}{Object of class {\link[base]{character}} expected. Well identification.}
}
\description{
Check if the item "well_id" have one unique class identified as character.
}
