% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_well_minus10_weigth}
\alias{check_well_minus10_weigth}
\title{Attribute "well_minus10_weigth" verification}
\usage{
check_well_minus10_weigth(well_minus10_weigth)
}
\arguments{
\item{well_minus10_weigth}{Object of class {\link[base]{integer}} expected. Catch weight of individus less than 10 tonnes (by well, in tonne, all species considerated).}
}
\description{
Check if the item "well_minus10_weigth" have one unique class identified as integer and if value are >= 0.
}
