% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_wellplan_id}
\alias{check_wellplan_id}
\title{Attribute "wellplan_id" verification}
\usage{
check_wellplan_id(wellplan_id)
}
\arguments{
\item{wellplan_id}{Object of class {\link[base]{character}} expected. Wellplan identification.}
}
\description{
Check if the item "wellplan_id" have one unique class identified as character.
}
