% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_wellplan_number}
\alias{check_wellplan_number}
\title{Attribut "wellplan_number" verification}
\usage{
check_wellplan_number(wellplan_number)
}
\arguments{
\item{wellplan_number}{(integer) Well plan number of individus.}
}
\description{
Check if the item "wellplan_number" have one unique class identified as integer.
}
