% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_initialisation.R
\name{data_model_initialisation}
\alias{data_model_initialisation}
\title{Data model initialisation}
\usage{
data_model_initialisation(
  periode_reference,
  countries,
  db_con,
  sample_type,
  log_file = FALSE,
  log_path = NULL,
  log_name = "data_model_initialisation"
)
}
\arguments{
\item{periode_reference}{(integer) Year(s) of the reference period coded on 4 digits.}

\item{countries}{(character) ISO code on 3 letters related to one or more countries.}

\item{db_con}{(PostgreSQLConnection) An R's object which contain connexion identifiers for a database.}

\item{log_file}{(logical) Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{(character) Path of the log file directory. By default NULL.}

\item{log_name}{(character) Name of the log file. By default "t3log".}
}
\value{
The function return two R6 reference object, one class "object_model_data" and the second class "object_full_trips".
}
\description{
Shortcut for initialisation of data's object model.
}
