% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2boot.R
\name{df2boot}
\alias{df2boot}
\title{Translate Data frame to boot class}
\usage{
df2boot(t0, t, R, sample_data, strata)
}
\arguments{
\item{t0}{The observed value of statistic applied to data}

\item{t}{A matrix with sum(R) rows each of which is a bootstrap replicate of the result of calling statistic.}

\item{R}{The number of bootstrap replicates. Usually this will be a single positive integer.}

\item{sample_data}{The sample data}

\item{strata}{An integer vector or factor specifying the strata for multi-sample problems. This may be specified for any simulation, but is ignored when sim = "parametric". When strata is supplied for a nonparametric bootstrap, the simulations are done within the specified strata}
}
\description{
Function to translate data frame to boot class (boot package) for ci computation
}
