% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarycatch.R
\name{elementarycatch}
\alias{elementarycatch}
\title{R6 class elementarycatch}
\description{
Create R6 reference object class elementarycatch
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarycatch-new}{\code{elementarycatch$new()}}
\item \href{#method-elementarycatch-clone}{\code{elementarycatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 activities class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatch$new(
  activity_id,
  elementarycatch_id,
  ocean,
  school_type,
  logbook_category,
  logbook_category_name,
  specie_code,
  specie_code3l,
  catch_weight
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity_id}}{Object of class {\link[base]{character}} expected. Activity identification.}

\item{\code{elementarycatch_id}}{Object of class {\link[base]{character}} expected. Elementary catch identification.}

\item{\code{ocean}}{Object of class {\link[base]{integer}} expected. Ocean identification.}

\item{\code{school_type}}{Object of class {\link[base]{integer}} expected. School type identification.}

\item{\code{logbook_category}}{Object of class {\link[base]{integer}} expected. Logbook weight category.}

\item{\code{logbook_category_name}}{Object of class {\link[base]{character}} expected. Logbook category name identification.}

\item{\code{specie_code}}{Object of class {\link[base]{integer}} expected. Specie code identification.}

\item{\code{specie_code3l}}{Object of class {\link[base]{character}} expected. Specie code identification on 3 characters.}

\item{\code{catch_weight}}{Object of class {\link[base]{numeric}} expected. Catch weight in tonnes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
