% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarycatches.R
\docType{data}
\name{elementarycatches}
\alias{elementarycatches}
\title{R6 class elementarycatches creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
elementarycatches
}
\description{
Create R6 reference object class elementarycatches
}
\keyword{datasets}
