% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarylandings.R
\name{elementarylandings}
\alias{elementarylandings}
\title{R6 class elementarylandings}
\description{
Create R6 reference object class elementarylandings
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{elementarylandings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarylandings-new}{\code{elementarylandings$new()}}
\item \href{#method-elementarylandings-add}{\code{elementarylandings$add()}}
\item \href{#method-elementarylandings-filter_by_trip}{\code{elementarylandings$filter_by_trip()}}
\item \href{#method-elementarylandings-clone}{\code{elementarylandings$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count"><a href='../../t3/html/list_t3.html#method-list_t3-count'><code>t3::list_t3$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract"><a href='../../t3/html/list_t3.html#method-list_t3-extract'><code>t3::list_t3$extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract_l1_element_value"><a href='../../t3/html/list_t3.html#method-list_t3-extract_l1_element_value'><code>t3::list_t3$extract_l1_element_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter"><a href='../../t3/html/list_t3.html#method-list_t3-filter'><code>t3::list_t3$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter_l1"><a href='../../t3/html/list_t3.html#method-list_t3-filter_l1'><code>t3::list_t3$filter_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="modification_l1"><a href='../../t3/html/list_t3.html#method-list_t3-modification_l1'><code>t3::list_t3$modification_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name"><a href='../../t3/html/list_t3.html#method-list_t3-name'><code>t3::list_t3$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove"><a href='../../t3/html/list_t3.html#method-list_t3-remove'><code>t3::list_t3$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view"><a href='../../t3/html/list_t3.html#method-list_t3-view'><code>t3::list_t3$view()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarylandings-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarylandings-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarylandings class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(empty, list or R6-elementarylanding classes) Nothing, a list of object R6-elementarylanding classes or one object R6-elementarylanding classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarylandings-add"></a>}}
\if{latex}{\out{\hypertarget{method-elementarylandings-add}{}}}
\subsection{Method \code{add()}}{
Function for add a new elementarylanding in the object elementarylandings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$add(new_item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_item}}{(list or R6-elementarylanding classes) A list of object R6-elementarylanding classes or one object R6-elementarylanding classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarylandings-filter_by_trip"></a>}}
\if{latex}{\out{\hypertarget{method-elementarylandings-filter_by_trip}{}}}
\subsection{Method \code{filter_by_trip()}}{
Function for filter elementarylandings by trip identification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$filter_by_trip(trip_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarylandings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarylandings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
