% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarysampleraw.R
\name{elementarysampleraw}
\alias{elementarysampleraw}
\title{R6 class trip}
\description{
Create R6 reference object class elementarysampleraw
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarysampleraw-new}{\code{elementarysampleraw$new()}}
\item \href{#method-elementarysampleraw-clone}{\code{elementarysampleraw$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarysampleraw-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarysampleraw-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarysampleraw class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$new(
  trip_id,
  well_id,
  well_id_source,
  sample_id,
  sub_sample_id,
  sub_sample_total_count_id,
  elementarysampleraw_id,
  sample_quality_code,
  sample_type_code,
  species_fao_code,
  size_measure_type_code,
  sample_total_count,
  sample_number_measured,
  sample_length_class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. Trip identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.
For the Observe database, the topiaid field from the table ps_logbook.well of the database is used;
if it is missing, the well ID is generated by the t3 process, as follows: well_id="fr.ird.data.ps.logbook.Well#trip_id#well_label".
For the AVDTH database, well identification is always generated by T3 process such as: well_id="fr.ird.avdth.entities.data.Well#vessel_code#trip_end_date#N_CUVE.F_POS_CUVE" .}

\item{\code{well_id_source}}{Object of class {\link[base]{character}} expected. Source of the well identification: "Observe" if the topiaid field from Observe database is used or "T3" if the the well ID is generated by the t3 process.}

\item{\code{sample_id}}{Object of class {\link[base]{character}} expected. Sample identification.}

\item{\code{sub_sample_id}}{Object of class {\link[base]{integer}} expected. Sub sample identification.}

\item{\code{sub_sample_total_count_id}}{Object of class {\link[base]{integer}} expected. Sub sample identification bis in relation with the fish total count.}

\item{\code{elementarysampleraw_id}}{Object of class {\link[base]{character}} expected. Elementary sample raw identification.}

\item{\code{sample_quality_code}}{Object of class {\link[base]{integer}} expected. Sample quality identification.}

\item{\code{sample_type_code}}{Object of class {\link[base]{integer}} expected. Sample type identification.}

\item{\code{species_fao_code}}{Object of class {\link[base]{character}} expected. Specie code identification on 3 characters.}

\item{\code{size_measure_type_code}}{Object of class {\link[base]{integer}} expected. Length type identification.}

\item{\code{sample_total_count}}{Object of class {\link[base]{integer}} expected. Sample number of total individuals counted.}

\item{\code{sample_number_measured}}{Object of class {\link[base]{integer}} expected. Sample number of measured individuals.}

\item{\code{sample_length_class}}{Object of class {\link[base]{double}} expected. Sample length class of measured individuals.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarysampleraw-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarysampleraw-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
