% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R
\docType{data}
\name{full_trips}
\alias{full_trips}
\title{R6 class full_trips creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
full_trips
}
\description{
Create R6 reference object class full_trips
}
\keyword{datasets}
