% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_directory.R
\name{initiate_directory}
\alias{initiate_directory}
\title{Output directories creation}
\usage{
initiate_directory(output_path, new_directory = TRUE, level = "all")
}
\arguments{
\item{output_path}{Object of class \code{\link[base]{character}} expected. Output path directory.}

\item{new_directory}{Object of class \code{\link[base]{logical}} expected. Initiate a new output directory of use an existing one. By default TRUE.}

\item{level}{Object of class \code{\link[base]{character}} expected. Initiation of a full output directory ("all") or specific to a level (choose between "level1", "level2", "until_level2 or "level3").}
}
\value{
A list with output directory path, at least, and log path if requested.
}
\description{
Initiation of output directories creation.
}
