% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_log_file.R
\name{initiate_log_file}
\alias{initiate_log_file}
\title{Log file creation}
\usage{
initiate_log_file(log_path)
}
\arguments{
\item{log_path}{(character) Path of the log file.}
}
\value{
A log file in text format at the location define in the argument "log_path".
}
\description{
Initiation a log file creation.
}
