% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{landing_weight}
\alias{landing_weight}
\alias{check_landing_weight}
\title{Attribut "landing_weight" verification}
\usage{
check_landing_weight(landing_weight)
}
\arguments{
\item{landing_weight}{(numeric) Landing weight in tonne}
}
\description{
Check if the item "landing_weight" have one unique class identified as numeric
}
