% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_weight_relationship}
\alias{length_weight_relationship}
\title{Referential LWR (length weight relationship) table}
\format{
length_weight_relationship: data frame with 40 rows and 7 columns:
\describe{
  \item{ocean_code}{Ocean code, type \code{\link[base]{integer}}. For example \code{ocean_code=1} for the Atlantic Ocean and \code{ocean_code=2} the Indian Ocean}
  \item{ocean_label}{Ocean label ("Atlantic" or "Indian"), type \code{\link[base]{character}}.}
  \item{species_code}{Species code, type \code{\link[base]{integer}}.}
  \item{species_fao_code}{Species FAO code, type \code{\link[base]{character}}.}
  \item{length_weight_formula}{LWR formula form, type \code{\link[base]{character}}.}
  \item{lwr_a}{Parameter a of LWR formula, type \code{\link[base]{double}}.}
  \item{lwr_b}{Parameter b of LWR formula, type \code{\link[base]{double}}.}
}
}
\usage{
data(length_weight_relationship, package="t3")
}
\description{
Conversion of the samples length measurements in weight by length weight relationship (LWR).\cr
Formula associated are the following one:\cr
\eqn{RWT=a \times LF^b}{RWT=a x LF^b} where:
\describe{
  \item{RWT}{is the round weight (kg)}
  \item{LF }{is the curved fork length (cm)}
  \item{a, b}{parameters from the reference table, dependent of of the species and potentially of the area (ocean or others) and the season.}
}
}
\details{
More detail information could be find on the regional fisheries management organisations (RFMOs) like \href{https://www.iccat.int/}{ICCAT} or \href{https://iotc.org/}{IOTC}.
}
\keyword{length}
\keyword{relationship}
\keyword{weight}
