% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_trip.R
\name{obj_trip}
\alias{obj_trip}
\title{Objet trip creation}
\usage{
obj_trip(data)
}
\arguments{
\item{data}{A data frame or a list with data in trip table format of a T3 database. See detail section for more informations.}
}
\description{
Generate objet trip with S4 method (from T3 database format).
}
\details{
To do after dude !
}
\references{
\url{https://github.com/OB7-IRD/t3}
}
\author{
Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr}
}
