% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_activity.R
\name{object_activity}
\alias{object_activity}
\title{R6 object object_activity creation}
\usage{
object_activity(trip_id, activity_id, ocean, activity_date,
  activity_number, set_count, school_type, activity_code, activity_name,
  time_at_sea)
}
\value{
A R6 reference object with data related to one activity
}
\description{
Creation object object_activity in relation with R6 reference object class activity
}
\seealso{
\code{\link{activity}}
}
