% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarycatches.R
\name{object_elementarycatches}
\alias{object_elementarycatches}
\title{R6 object object_elementarycatches creation}
\usage{
object_elementarycatches(...)
}
\value{
A R6 reference object with data related to one or several elementary catches
}
\description{
Creation object object_elementarycatches in relation with R6 reference object class elementarycatches
}
\seealso{
\code{\link{elementarycatches}}
}
