% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarylanding.R
\name{object_elementarylanding}
\alias{object_elementarylanding}
\title{R6 object object_elementarylanding creation}
\usage{
object_elementarylanding(
  trip_id,
  landing_category,
  landing_category_name,
  specie_code3l,
  landing_weight
)
}
\value{
A R6 reference object with data related to one elementary landing
}
\description{
Creation object object_elementarylanding in relation with R6 reference object class elementarylanding
}
\seealso{
\code{\link{elementarylanding}}
}
