% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarylandings.R
\name{object_elementarylandings}
\alias{object_elementarylandings}
\title{R6 object object_elementarylandings creation}
\usage{
object_elementarylandings(...)
}
\value{
A R6 reference object with data related to one or several elementary landing(s)
}
\description{
Creation object object_elementarylandings in relation with R6 reference object class elementarylandings
}
\seealso{
\code{\link{elementarylandings}}
}
