% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarysample.R
\name{object_elementarysample}
\alias{object_elementarysample}
\title{R6 object object_elementarysample creation}
\usage{
object_elementarysample(
  trip_id,
  sample_id,
  sub_sample_id,
  specie_code3l,
  length_type,
  sample_total_count,
  sample_number_measured,
  sample_length_class
)
}
\value{
A R6 reference object with data related to one elementary sample
}
\description{
Creation object object_elementarysample in relation with R6 reference object class elementarysample
}
\seealso{
\code{\link{elementarysample}}
}
