% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarysampleraw.R
\name{object_elementarysampleraw}
\alias{object_elementarysampleraw}
\title{R6 object object_elementarysampleraw creation}
\usage{
object_elementarysampleraw(
  trip_id,
  sample_id,
  sub_sample_id,
  specie_code3l,
  length_type,
  sample_total_count,
  sample_number_measured,
  sample_length_class
)
}
\value{
A R6 reference object with data related to one elementary sample (raw data)
}
\description{
Creation object object_elementarysampleraw in relation with R6 reference object class elementarysampleraw
}
\seealso{
\code{\link{elementarysampleraw}}
}
