% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarywellplan.R
\name{object_elementarywellplan}
\alias{object_elementarywellplan}
\title{R6 object object_elementarywellplan creation}
\usage{
object_elementarywellplan(
  wellplan_id,
  well_id,
  activity_id,
  specie_code3l,
  wellplan_weight,
  wellplan_number,
  wellplan_weigth_category_code,
  wellplan_weigth_category_label
)
}
\value{
A R6 reference object with data related to one item in the well plan
}
\description{
Creation object object_elementarywellplan in relation with R6 reference object class elementarywellplan
}
\seealso{
\code{\link{elementarywellplan}}
}
