% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_full_trips.R
\name{object_full_trips}
\alias{object_full_trips}
\title{R6 object full trips creation}
\usage{
object_full_trips(object_trips)
}
\arguments{
\item{object_trips}{(class R6 and trips) A R6 reference object of class trips}
}
\value{
A R6 reference object with data of trips grouped by full trips
}
\description{
Creation object full trips in relation with R6 reference object class full_trips
}
\seealso{
\code{\link{full_trips}}
}
