% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_r6.R
\name{object_r6}
\alias{object_r6}
\title{R6 object creation}
\usage{
object_r6(class_name, ...)
}
\arguments{
\item{class_name}{(character) Name of the R6 reference class}

\item{...}{(empty, list or R6 object) Leave empty if you want to initialise the object or provide one or several item to add in the class}
}
\value{
A R6 reference object
}
\description{
Creation object R6 in relation with R6 reference object class declared in function argument
}
