% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_trips.R
\name{object_trips}
\alias{object_trips}
\title{R6 object trips creation}
\usage{
object_trips(...)
}
\arguments{
\item{...}{(empty or class R6 and trip) Leave empty if you want to initialise the object or provide at least one R6 reference object of class trip}
}
\value{
A R6 reference object with data of one or more trips
}
\description{
Creation object trips in relation with R6 reference object class trips
}
\seealso{
\code{\link{trips}}
}
