% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_well.R
\name{object_well}
\alias{object_well}
\title{R6 object object_well creation}
\usage{
object_well(
  well_id,
  well_minus10_weigth,
  well_plus10_weigth,
  well_global_weigth
)
}
\value{
A R6 reference object with data related to one well
}
\description{
Creation object object_well in relation with R6 reference object class well
}
\seealso{
\code{\link{well}}
}
