% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observe_logbook_program}
\alias{observe_logbook_program}
\title{Reference table defining the logbook program for Observe database to be considered in T3 process.}
\format{
observe_logbook_program: data frame with 29 rows and 3 columns:
\itemize{
\item{topiaid: } logbook program topiaid, type \code{\link[base]{character}}.
\item{code: }  program code, type \code{\link[base]{character}}.
\item{label1: } program label in English, type \code{\link[base]{numeric}}.
}
}
\usage{
data(observe_logbook_program, package="t3")
}
\description{
Reference table defining the logbook program for Observe database to be considered in T3 process.
}
\examples{
data(observe_logbook_program, package="t3")
 observe_logbook_program
}
\keyword{wobserve_logbook_program}
