% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{sample_number_measured}
\alias{sample_number_measured}
\alias{check_sample_number_measured}
\title{Attribut "sample_number_measured" verification}
\usage{
check_sample_number_measured(sample_number_measured)
}
\arguments{
\item{sample_number_measured}{(integer) Sample number of measured individus}
}
\description{
Check if the item "sample_number_measured" have one unique class identified as integer and if all values are positive and different of zero
}
