% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{set_duration_ref}
\alias{set_duration_ref}
\title{Referential **set duration** table}
\format{
## **`set_duration_ref`**:
A data frame with 819 rows and 9 columns:
\describe{
 \item{year}{Year, type \code{\link[base]{integer}}.}
  \item{flag_code}{Country code, type \code{\link[base]{integer}}.}
  \item{flag_code_iso_3}{3 letter ISO country codes, type \code{\link[base]{character}}.}
  \item{ocean_code}{Ocean code, type \code{\link[base]{integer}}. For example \code{ocean_code=1} for the Atlantic Ocean and \code{ocean_code=2} the Indian Ocean}
  \item{school_type_code_avdth}{School type code in AVDTH referential template, type \code{\link[base]{integer}}, (1 for floating object school, 2 for free school and 3 for undetermined school).}
  \item{school_type_code_observe}{School type code in Observe referential template, type \code{\link[base]{integer}}, (1 for floating object school, 2 for free school and 0 for undetermined school).}
  \item{parameter_a}{Slope of the linear function giving the duration of the set as a function of the catch weight, type \code{\link[base]{double}}.}
  \item{parameter_b}{Intercept of the linear function giving the duration of the set as a function of the catch weight, type \code{\link[base]{double}}.}
  \item{null_set_value}{Specific duration value (min) provided for a null set (a fishing set without any catches associated), type \code{\link[base]{double}}.}
}
}
\usage{
data(set_duration_ref, package="t3")
}
\description{
The set duration (min) is calculated from catch weight (t), according to a linear function with two parameters a and b:
\eqn{set_duration = a \times catch_weight + b}{set_duration = a x catch_weight + b}.
These are found through this reference table, for each year, ocean, fishing school and country.
Furthermore, a specific value provided for a null set (a fishing set without any catches associated).
}
\keyword{datasets}
