% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisedsample.R
\name{standardisedsample}
\alias{standardisedsample}
\title{R6 class standardisedsample}
\description{
Create R6 reference object class standardisedsample
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-standardisedsample-new}{\code{standardisedsample$new()}}
\item \href{#method-standardisedsample-clone}{\code{standardisedsample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardisedsample-new"></a>}}
\if{latex}{\out{\hypertarget{method-standardisedsample-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 standardisedsample class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$new(
  trip_id,
  well_id,
  well_id_source,
  sample_id,
  sample_quality_code,
  sample_type_code,
  species_fao_code,
  sample_standardised_length_class_lf,
  sample_number_measured_extrapolated_lf,
  sample_total_count
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.
For the Observe database, the topiaid field from the table ps_logbook.well of the database is used;
if it is missing, the well ID is generated by the t3 process, as follows: well_id="fr.ird.data.ps.logbook.Well#trip_id#well_label".
For the AVDTH database, well identification is always generated by T3 process such as: well_id="fr.ird.avdth.entities.data.Well#vessel_code#trip_end_date#N_CUVE.F_POS_CUVE" .}

\item{\code{well_id_source}}{Object of class {\link[base]{character}} expected. Source of the well identification: "Observe" if the topiaid field from Observe database is used or "T3" if the the well ID is generated by the t3 process.}

\item{\code{sample_id}}{Object of class {\link[base]{character}} expected. Sample identification.}

\item{\code{sample_quality_code}}{Object of class {\link[base]{integer}} expected. Sample quality identification.}

\item{\code{sample_type_code}}{Object of class {\link[base]{integer}} expected. Sample type identification.}

\item{\code{species_fao_code}}{(character) Specie code identification on 3 characters.}

\item{\code{sample_standardised_length_class_lf}}{Object of class {\link[base]{integer}} expected. Sample standardized length class length fork of measured individuals.}

\item{\code{sample_number_measured_extrapolated_lf}}{Object of class {\link[base]{numeric}} expected. Sample number of measured individuals extrapolated to all counted individuals.}

\item{\code{sample_total_count}}{Object of class {\link[base]{integer}} expected. Sample number of total individuals counted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardisedsample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-standardisedsample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
