% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3-package.R
\docType{package}
\name{t3-package}
\alias{t3}
\alias{t3-package}
\title{Tropical Tunas Treatment (T3)}
\description{
The T3 processing was built about 30 years ago in order to correct biases of the logbook data on species composition and to provide more accurate catch estimates per species for the European purse seine fleet.
}
\section{R6 objets model creation}{

\itemize{
\item{\code{\link{object_model_data}}}
}
}

\section{Methods}{

\itemize{
\item{\code{\link{full_trips}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/t3}
  \item Report bugs at \url{https://github.com/OB7-IRD/t3/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr}

Authors:
\itemize{
  \item name_Bob1 surname_bob1 \email{bob1@ird.fr}
}

Other contributors:
\itemize{
  \item name_Bob2 surname_bob2 \email{bob2@ird.fr} [copyright holder]
}

}
\keyword{internal}
