% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_level2.R
\name{t3_level2}
\alias{t3_level2}
\title{T3 process level 2}
\usage{
t3_level2(
  length_step,
  sample_set,
  length_weight_relationship_data,
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level2"
)
}
\arguments{
\item{length_step}{(data.frame) Data frame object with length ratio between ld1 and lf class.}

\item{sample_set}{(data.frame) Data frame object with weighted weigth of each set sampled.}

\item{length_weight_relationship_data}{(data.frame) Data frame object with parameters for length weight relationship.}

\item{log_file}{(logical) Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{(character) Path of the log file directory. By default NULL.}

\item{log_name}{(character) Name of the log file. By default "t3_level2".}
}
\value{
The function a R6 reference object of class "object_full_trips".
}
\description{
Level 2 of t3 process (samples standardisation).
}
