% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_level3.R
\name{t3_level3}
\alias{t3_level3}
\title{T3 process level 3}
\usage{
t3_level3(
  inputs_level3,
  inputs_level3_path = NULL,
  periode_reference_level3 = NULL,
  target_year,
  target_ocean = NULL,
  period_duration,
  country_flag = NULL,
  input_type = "observe_database",
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = FALSE,
  plot_sample = FALSE,
  avdth_patch_coord = FALSE,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE,
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level3",
  output_path = NULL,
  new_directory = FALSE,
  integrated_process = FALSE
)
}
\arguments{
\item{inputs_level3}{Object of type \code{\link[base]{data.frame}} expected. Inputs of levels 3 (see function path to level 3).}

\item{inputs_level3_path}{Object of type \code{\link[base]{character}} expected. Path to the folder containing yearly data output of the level 1 and 2 (output of the function the path to level 3). If provide, replace the inputs_level3 object.}

\item{periode_reference_level3}{Object of type \code{\link[base]{integer}} expected. Year(s) period of reference for modelling estimation.}

\item{target_year}{Object of type \code{\link[base]{integer}} expected. Year of interest for the model estimation and prediction.Default value is current year -1.}

\item{target_ocean}{Object of type \code{\link[base]{integer}} expected. The code of ocean of interest.}

\item{period_duration}{Object of type \code{\link[base]{integer}} expected. number of years use for the modelling. The default value is 5}

\item{country_flag}{Three letters FAO flag code of country to estimate catches.}

\item{input_type}{Type of coding use in different databases. Default value is 'observe_database'. Values can be 'observe_database' or 'avdth_database'.}

\item{distance_maximum}{Object of type \code{\link[base]{integer}} expected. Maximum distance between all sets of a sampled well. By default 5.}

\item{number_sets_maximum}{Object of type \code{\link[base]{integer}} expected. Maximum number of sets allowed in mixture. By default 5.}

\item{set_weight_minimum}{Object of type \code{\link[base]{integer}} expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.}

\item{minimum_set_frequency}{Object of type \code{\link[base]{numeric}} expected. Minimum threshold proportion of set in awell to be used for model training in the process. By default 0.1.}

\item{vessel_id_ignored}{Object of type \code{\link[base]{integer}} expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.}

\item{num.trees}{Object of type \code{\link[base]{integer}} expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.}

\item{mtry}{Object of type \code{\link[base]{integer}} expected. Number of variables randomly sampled as candidates at each split. The default value is 2.}

\item{min.node.size}{Object of type \code{\link[base]{numeric}} expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.}

\item{seed_number}{Object of type \code{\link[base]{integer}} expected. Set the initial seed for the modelling. The default value is 7.}

\item{small_fish_only}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}

\item{plot_sample}{\code{\link[base]{logical}}. Whether the sample figure is computed. Default value = F}

\item{avdth_patch_coord}{parameter waiting for coordinate conversion patch from avdth database}

\item{ci}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.}

\item{ci_type}{Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if  the ci parameter is TRUE.}

\item{Nboot}{Object of type \code{\link[base]{numeric}} expected. The number of bootstrap samples desired for the ci computation. The default value is 10.}

\item{plot_predict}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether maps of catch at size have to be done.}

\item{log_file}{Object of type \code{\link[base]{logical}} expected. Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{Object of type \code{\link[base]{character}} expected. Path of the log file directory. By default NULL.}

\item{log_name}{Object of type \code{\link[base]{character}} expected. Name of the log file. By default "t3_level2".}

\item{output_path}{Object of class \code{\link[base]{character}} expected. Outputs path directory. By default NULL.}

\item{new_directory}{Object of class \code{\link[base]{logical}} expected. Initiate a new outputs directory of use an existing one. By default NULL.}

\item{integrated_process}{Object of class \code{\link[base]{logical}} expected. Indicate if the process is integrated in another (like the one in the function "t3_process"). By default FALSE.}
}
\description{
Level 3 of t3 process (modelislation).
}
