% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_process.R
\name{t3_process}
\alias{t3_process}
\title{T3 process compilation}
\usage{
t3_process(
  process = "all",
  log_file = FALSE,
  log_path = NULL,
  db_con,
  periode_reference,
  countries,
  oceans,
  sample_type,
  species_rf1,
  set_duration_ref,
  length_step,
  sample_set,
  length_weight_relationship_data,
  trips_selected = NULL
)
}
\arguments{
\item{process}{(character) Specify here if you want to run the whole process or just a part of it. By default "all". Check detail section below for more information.}

\item{log_file}{(logical) Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{(character) Path of the log file directory. By default NULL.}

\item{db_con}{(PostgreSQLConnection) An R's object which contain connexion identifiers for a t3 database.}

\item{periode_reference}{(integer) Year(s) of the reference period coded on 4 digits.}

\item{countries}{(character) ISO code on 3 letters related to one or more countries.}

\item{oceans}{(integer) Ocean(s) related to data coded on 1 digit.}

\item{species_rf1}{(integer) Specie(s) code(s) used for the rf1 process.}

\item{set_duration_ref}{(data frame) Data and parameters for set duration calculation (by year, country, ocean and school type).}

\item{length_step}{(data.frame) Data frame object with length ratio between ld1 and lf class.}

\item{sample_set}{(data.frame) Data frame object with weighted weigth of each set sampled.}

\item{length_weight_relationship_data}{(data.frame) Data frame object with parameters for length weight relationship.}

\item{trips_selected}{(character) Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference, countries and sample types). By default NULL.}
}
\value{
The function a R6 reference object of class "object_full_trips".
}
\description{
Run the t3 process, with the possibility to run all the process or stop.
}
\details{
For the argument "process", you can choose between 5 modalities (descending size classification):
\itemize{
 \item{all: }{argument by default, you launch all the process}
 \item{level1: }{you launch data model initialisation and the process level 1}
 \item{level2: }{you launch data model initialisation and the process level 2}
 \item{until_level2: }{you launch data model initialisation, the process level 1 and 2}
}
}
