% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_t3.R
\docType{data}
\name{tools_t3}
\alias{tools_t3}
\title{R6 class tools_t3 creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tools_t3
}
\description{
Create R6 reference object class tools_t3
}
\keyword{datasets}
