% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunaboot.R
\name{tunaboot}
\alias{tunaboot}
\title{Boostrap for Model Predictions}
\usage{
tunaboot(
  sample_data,
  allset_data,
  Ntree = 1000,
  Nmtry = 2,
  Nseed = 7,
  min_node = 5,
  Nboot = 10,
  Nseed_boot,
  target_period
)
}
\arguments{
\item{sample_data}{(data frame) Data used for the modelling. Output table from process 3.1.}

\item{allset_data}{(data frame) Data used for prediction. Output table from process 3.4.}

\item{Ntree}{(integer) Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.}

\item{Nmtry}{Number of variables randomly sampled as candidates at each split. The default value is 2.}

\item{Nseed}{Set the initial seed for the modelling. The default value is 7.}

\item{min_node}{Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.}

\item{Nboot}{The number of bootstrap samples desired. The default value is 10}

\item{Nseed_boot}{Set the initial seed for the modelling. The default value is equal to Nseed parameter}

\item{target_period}{Time period for the predictions in year. Default is the year of the data to predict}
}
\description{
Bootstrap function which compute proportion by species fishing mode and ocean for confindence interval
}
