% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{R6 class activity creation}
\description{
Create R6 reference object class activity
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{activity$new()}}
\item \href{#method-clone}{\code{activity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{activity$new(
  trip_id,
  activity_id,
  ocean,
  activity_date,
  activity_number,
  activity_latitude,
  activity_longitude,
  set_count,
  school_type,
  activity_code,
  activity_name,
  time_at_sea
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{activity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
