% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{R6 class activity}
\description{
Create R6 reference object class activity
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-activity-new}{\code{activity$new()}}
\item \href{#method-activity-clone}{\code{activity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activity-new"></a>}}
\if{latex}{\out{\hypertarget{method-activity-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 activities class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{activity$new(
  trip_id,
  activity_id,
  ocean_code,
  activity_date,
  activity_number,
  activity_latitude,
  activity_longitude,
  set_count,
  set_success_status_code,
  set_success_status_label,
  school_type_code,
  activity_code,
  activity_label,
  objectoperation_code,
  objectoperation_label,
  objectoperation_id,
  time_at_sea
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. Trip identification (unique topiaid from database).}

\item{\code{activity_id}}{Object of class {\link[base]{character}} expected. Activity identification (unique topiaid from database).}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. Ocean identification.}

\item{\code{activity_date}}{Object of class {\link[base]{character}} expected. Activity date in format year month day.}

\item{\code{activity_number}}{Object of class {\link[base]{integer}} expected. Activity number.}

\item{\code{activity_latitude}}{Object of class {\link[base]{numeric}} expected. Latitude, in decimal degree, of the activity.}

\item{\code{activity_longitude}}{Object of class {\link[base]{numeric}} expected. Longitude, in decimal degree, of the activity.}

\item{\code{set_count}}{Object of class {\link[base]{integer}} expected. Number of set associated to the activity.}

\item{\code{set_success_status_code}}{Object of class {\link[base]{integer}} expected. Code status of the set.}

\item{\code{set_success_status_label}}{Object of class {\link[base]{character}} expected. Label status of the set.}

\item{\code{school_type_code}}{Object of class {\link[base]{integer}} expected. School type identification.}

\item{\code{activity_code}}{Object of class {\link[base]{integer}} expected. Activity code identification.}

\item{\code{activity_label}}{Object of class {\link[base]{character}} expected. Activity identification.}

\item{\code{objectoperation_code}}{Object of class {\link[base]{integer}} expected. Operation on floating object code identification.}

\item{\code{objectoperation_label}}{Object of class {\link[base]{character}} expected. Operation on floating object identification.}

\item{\code{objectoperation_id}}{Object of class {\link[base]{character}} expected. Operation on floating object identification (unique topiaid from database).}

\item{\code{time_at_sea}}{Object of class {\link[base]{integer}} expected. Time at sea in hours.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-activity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{activity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
