% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_ci_calculator.R
\name{catch_ci_calculator}
\alias{catch_ci_calculator}
\title{Compute final confidence interval}
\usage{
catch_ci_calculator(fit_data, boot_data)
}
\arguments{
\item{fit_data}{standardized catch estimates from the prediction step}

\item{boot_data}{standardized bootstrap catch data from the bootstap step}
}
\value{
A vector of fitted data and confidence interval
fit_data Fitted data from the input
ci_inf Estimated inferior value of the confidence interval
ci_sup Estimated superior value of the confidence interval
}
\description{
Compute final confidence interval
}
