% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_elementarylanding_id}
\alias{check_elementarylanding_id}
\title{Attribut "elementarylanding_id" verification}
\usage{
check_elementarylanding_id(elementarylanding_id)
}
\arguments{
\item{elementarylanding_id}{(character) Elementary landing identification}
}
\description{
Check if the item "elementarylanding_id" have one unique class identified as character
}
