% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fleet}
\alias{check_fleet}
\title{Attribut "fleet" verification}
\usage{
check_fleet(fleet)
}
\arguments{
\item{fleet}{(character) Fleet identification}
}
\description{
Check if the item "fleet" have one unique class identified as character
}
