% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_landing_date}
\alias{check_landing_date}
\title{Attribut "landing_date" verification}
\usage{
check_landing_date(landing_date)
}
\arguments{
\item{landing_date}{(date and hours) Landing date in format ymd_hms UTC}
}
\description{
Check if the item "landing_date" is in date format ymd_hms and check if the value is inferior to the actual date
}
