% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_specie_code}
\alias{check_specie_code}
\title{Attribute "specie_code" verification}
\usage{
check_specie_code(specie_code)
}
\arguments{
\item{specie_code}{Object of class {\link[base]{integer}} expected. Specie code identification.}
}
\description{
Check if the item "specie_code" have one unique class identified as integer and superior to zero.
}
