% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_specie_code3l}
\alias{check_specie_code3l}
\title{Attribut "specie_code3l" verification}
\usage{
check_specie_code3l(specie_code3l)
}
\arguments{
\item{specie_code3l}{(character) Specie code identification on 3 characters}
}
\description{
Check if the item "specie_code3l" have one unique class identified as character and 3 characters
}
