% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sub_sample_id}
\alias{check_sub_sample_id}
\title{Attribute "sub_sample_id" verification}
\usage{
check_sub_sample_id(sub_sample_id)
}
\arguments{
\item{sub_sample_id}{Object of class {\link[base]{integer}} expected. Sub sample identification.}
}
\description{
Check if the item "sub_sample_id" have one unique class identified as integer.
}
