% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_time_at_sea}
\alias{check_time_at_sea}
\title{Attribut "time_at_sea" verification}
\usage{
check_time_at_sea(time_at_sea)
}
\arguments{
\item{time_at_sea}{(integer) Time at sea in hours.}
}
\description{
Check if the item "time_at_sea" have one unique class identified as integer and if all values are superior or egal to zero.
}
