% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_trip_id}
\alias{check_trip_id}
\title{Attribut "trip_id" verification}
\usage{
check_trip_id(wellplan_id)
}
\arguments{
\item{trip_id}{(character) Trip identification.}
}
\description{
Check if the item "trip_id" have one unique class identified as character.
}
