% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vessel_id}
\alias{check_vessel_id}
\title{Attribut "vessel_id" verification}
\usage{
check_vessel_id(vessel_id)
}
\arguments{
\item{vessel_id}{(integer) Vessel identification}
}
\description{
Check if the item "vessel_id" have one unique class identified as integer
}
