% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vessel_type}
\alias{check_vessel_type}
\title{Attribute "vessel_type" verification}
\usage{
check_vessel_type(vessel_type)
}
\arguments{
\item{vessel_type}{Object of class {\link[base]{character}} expected. Vessel type identification.}
}
\description{
Check if the item "vessel_type" have one unique class identified as character.
}
