% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_initialisation.R
\name{data_model_initialisation}
\alias{data_model_initialisation}
\title{Data model initialisation}
\usage{
data_model_initialisation(
  data_source = "observe_database",
  database_connection,
  log_file = FALSE,
  log_path = NULL,
  log_name = "data_model_initialisation",
  years_period,
  flag_codes,
  ocean_codes,
  vessel_type_codes,
  species_fate_codes,
  sample_type_codes,
  weight_categories_avdth_ref = NULL,
  trip_ids = NULL
)
}
\arguments{
\item{data_source}{Object of class {\link[base]{character}} expected. Identification of data source. By default "observe_database" but you can switch to "avdth_database".}

\item{database_connection}{Database connection, list of one or more R object(s) expected. By default NULL.
Mandatory argument for data source "observe_database" ("PostgreSQLConnection" R object), corresponding to the second element of the object returned by \href{https://ob7-ird.github.io/furdeb/reference/postgresql_dbconnection.html}{`furdeb::postgresql_dbconnection()`}.
Or mandatory argument for data source"avdth_database" ("JDBCConnection" R object) corresponding to the second element of the object returned by \href{https://ob7-ird.github.io/furdeb/reference/access_dbconnection.html}{`furdeb::access_dbconnection()`}.
For data source "observe_database", a list of "PostgreSQLConnection" R objects can be specified to query data from different observe databases.
For example, a list of two database connection arguments for "observe_main" and "observe_acquisition" can be specified to simultaneously import and process recent data from acquisition database, which has not yet been imported into the main database, and older data from the main database.}

\item{log_file}{Object of class {\link[base]{logical}} expected. By default FALSE. Initiation or not for log file creation.}

\item{log_path}{Object of class {\link[base]{character}} expected. By default NULL. Path of the log file directory.}

\item{log_name}{Object of class {\link[base]{character}} expected. By default "data_model_initialisation". Name of the log file.}

\item{years_period}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{flag_codes}{Object of class {\link[base]{character}} expected. By default NULL. Three letters country(ies) FAO code(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{ocean_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{vessel_type_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{species_fate_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Specie fate(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{sample_type_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Sample type identification.}

\item{weight_categories_avdth_ref}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Reference table defining the logbook weight categories with the following columns:
\itemize{
\item{avdth_weight_category_code: } Logbook weight category code from avdth database, type \code{\link[base]{character}}.
\item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
\item{weight_category_min: } lower limit of the weight category (kg), type \code{\link[base]{numeric}}.
\item{weight_category_max: } upper limit of the weight category (kg), type \code{\link[base]{numeric}}.
}
Mandatory for \code{data_source = "avdth_database"}.}

\item{trip_ids}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".}
}
\value{
The function return a list with two R6 reference object inside: one class "object_model_data" and the second class "object_full_trips".
}
\description{
Shortcut for initialisation of data's object model from a observe or a AVDTH database.
}
