% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_initialisation.R
\name{data_model_initialisation}
\alias{data_model_initialisation}
\title{Data model initialisation}
\usage{
data_model_initialisation(
  data_source = "observe_database",
  database_connection,
  log_file = FALSE,
  log_path = NULL,
  log_name = "data_model_initialisation",
  time_period,
  fleet_code,
  ocean_code,
  vessel_type_code,
  species_fate_code,
  sample_type_code,
  trip_id = NULL
)
}
\arguments{
\item{data_source}{Object of class {\link[base]{character}} expected. Identification of data source. By default "observe_database" but you can switch to "avdth_database".}

\item{database_connection}{Database connection R object expected.  Mandatory argument for data source "observe_database" and "avdth_database".}

\item{log_file}{Object of class {\link[base]{logical}} expected. By default FALSE. Initiation or not for log file creation.}

\item{log_path}{Object of class {\link[base]{character}} expected. By default NULL. Path of the log file directory.}

\item{log_name}{Object of class {\link[base]{character}} expected. By default "data_model_initialisation". Name of the log file.}

\item{time_period}{Object of class {\link[base]{integer}} expected. Year(s) of the reference time period coded on 4 digits.}

\item{fleet_code}{Object of class {\link[base]{character}} expected. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{ocean_code}{Object of class {\link[base]{integer}} expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{vessel_type_code}{Object of class {\link[base]{integer}} expected. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{species_fate_code}{Object of class {\link[base]{integer}} expected. Specie fate(s) related to data extraction.}

\item{sample_type_code}{Object of class {\link[base]{integer}} expected. Sample type identification.}

\item{trip_id}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}
}
\value{
The function return a list with two R6 reference object inside: one class "object_model_data" and the second class "object_full_trips".
}
\description{
Shortcut for initialisation of data's object model from a observe or a AVDTH database.
}
