% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarycatch.R
\docType{data}
\name{elementarycatch}
\alias{elementarycatch}
\title{R6 class elementarycatch creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
elementarycatch
}
\description{
Create R6 reference object class elementarycatch
}
\keyword{datasets}
