% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarycatch.R
\name{elementarycatch}
\alias{elementarycatch}
\title{R6 class elementarycatch creation}
\description{
Create R6 reference object class elementarycatch
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarycatch$new()}}
\item \href{#method-clone}{\code{elementarycatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatch$new(
  activity_id,
  ocean,
  school_type,
  logbook_category,
  logbook_category_name,
  specie_code3l,
  catch_weight
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
