% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarycatches.R
\name{elementarycatches}
\alias{elementarycatches}
\title{R6 class elementarycatches}
\description{
Create R6 reference object class elementarycatches
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{elementarycatches}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarycatches-new}{\code{elementarycatches$new()}}
\item \href{#method-elementarycatches-add}{\code{elementarycatches$add()}}
\item \href{#method-elementarycatches-filter_by_activity}{\code{elementarycatches$filter_by_activity()}}
\item \href{#method-elementarycatches-clone}{\code{elementarycatches$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count"><a href='../../t3/html/list_t3.html#method-list_t3-count'><code>t3::list_t3$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract"><a href='../../t3/html/list_t3.html#method-list_t3-extract'><code>t3::list_t3$extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract_l1_element_value"><a href='../../t3/html/list_t3.html#method-list_t3-extract_l1_element_value'><code>t3::list_t3$extract_l1_element_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter"><a href='../../t3/html/list_t3.html#method-list_t3-filter'><code>t3::list_t3$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter_l1"><a href='../../t3/html/list_t3.html#method-list_t3-filter_l1'><code>t3::list_t3$filter_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="modification_l1"><a href='../../t3/html/list_t3.html#method-list_t3-modification_l1'><code>t3::list_t3$modification_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name"><a href='../../t3/html/list_t3.html#method-list_t3-name'><code>t3::list_t3$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove"><a href='../../t3/html/list_t3.html#method-list_t3-remove'><code>t3::list_t3$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view"><a href='../../t3/html/list_t3.html#method-list_t3-view'><code>t3::list_t3$view()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatches-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatches-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarycatches class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatches$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(empty, list or R6-elementarycatch classes) Nothing, a list of object R6-elementarycatch classes or one object R6-elementarycatch classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatches-add"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatches-add}{}}}
\subsection{Method \code{add()}}{
Function for add a new elementarycatch in the object elementarycatches.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatches$add(new_item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_item}}{(list or R6-elementarycatch classes) A list of object R6-elementarycatch classes or one object R6-elementarycatch classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatches-filter_by_activity"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatches-filter_by_activity}{}}}
\subsection{Method \code{filter_by_activity()}}{
Function for filter elementarycatches by activity identification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatches$filter_by_activity(activity_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity_id}}{(character) Activity identification.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarycatches-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarycatches-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarycatches$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
