% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarylanding.R
\name{elementarylanding}
\alias{elementarylanding}
\title{R6 class elementarylanding creation}
\description{
Create R6 reference object class elementarylanding
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarylanding$new()}}
\item \href{#method-clone}{\code{elementarylanding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarylanding class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylanding$new(
  trip_id,
  elementarylanding_id,
  landing_category,
  landing_category_name,
  specie_code,
  specie_code3l,
  landing_weight
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{elementarylanding_id}}{(character) Elementary landing identification.}

\item{\code{landing_category}}{(integer) Landing category identification.}

\item{\code{landing_category_name}}{(character) Landing category name identification.}

\item{\code{specie_code}}{(integer) Specie code identification.}

\item{\code{specie_code3l}}{(character) Specie code identification on 3 characters.}

\item{\code{landing_weight}}{(numeric) Landing weight in tonnes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylanding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
