% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarylanding.R
\docType{data}
\name{elementarylanding}
\alias{elementarylanding}
\title{R6 class elementarylanding creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
elementarylanding
}
\description{
Create R6 reference object class elementarylanding
}
\keyword{datasets}
