% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarylandings.R
\docType{data}
\name{elementarylandings}
\alias{elementarylandings}
\title{R6 class elementarylandings creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
elementarylandings
}
\description{
Create R6 reference object class elementarylandings
}
\keyword{datasets}
