% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarylandings.R
\name{elementarylandings}
\alias{elementarylandings}
\title{R6 class elementarylandings creation}
\description{
Create R6 reference object class elementarylandings
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{elementarylandings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarylandings$new()}}
\item \href{#method-add}{\code{elementarylandings$add()}}
\item \href{#method-filter_by_trip}{\code{elementarylandings$filter_by_trip()}}
\item \href{#method-clone}{\code{elementarylandings$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count">}\href{../../t3/html/list_t3.html#method-count}{\code{t3::list_t3$count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract">}\href{../../t3/html/list_t3.html#method-extract}{\code{t3::list_t3$extract()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter">}\href{../../t3/html/list_t3.html#method-filter}{\code{t3::list_t3$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name">}\href{../../t3/html/list_t3.html#method-name}{\code{t3::list_t3$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove">}\href{../../t3/html/list_t3.html#method-remove}{\code{t3::list_t3$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view">}\href{../../t3/html/list_t3.html#method-view}{\code{t3::list_t3$view()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarylandings class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(empty, list or R6-elementarylanding classes) Nothing, a list of object R6-elementarylanding classes or one object R6-elementarylanding classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Function for add a new elementarylanding in the object elementarylandings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$add(new_item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_item}}{(list or R6-elementarylanding classes) A list of object R6-elementarylanding classes or one object R6-elementarylanding classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_by_trip"></a>}}
\if{latex}{\out{\hypertarget{method-filter_by_trip}{}}}
\subsection{Method \code{filter_by_trip()}}{
Function for filter elementarylandings by trip identification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$filter_by_trip(trip_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarylandings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
