% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarysample.R
\name{elementarysample}
\alias{elementarysample}
\title{R6 class trip}
\description{
Create R6 reference object class elementarysample
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarysample-new}{\code{elementarysample$new()}}
\item \href{#method-elementarysample-clone}{\code{elementarysample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarysample-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarysample-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarysample class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysample$new(
  trip_id,
  well_id,
  sample_id,
  sub_sample_id,
  sample_quality_code,
  sample_type_code,
  species_code,
  species_fao_code,
  sample_standardised_length_class_lf,
  sample_number_measured_extrapolated_lf,
  sample_total_count
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. Trip identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.}

\item{\code{sample_id}}{Object of class {\link[base]{character}} expected. Sample identification.}

\item{\code{sub_sample_id}}{Object of class {\link[base]{integer}} expected. Sub sample identification.}

\item{\code{sample_quality_code}}{Object of class {\link[base]{integer}} expected. Sample quality identification.}

\item{\code{sample_type_code}}{Object of class {\link[base]{integer}} expected. Sample type identification.}

\item{\code{species_code}}{Object of class {\link[base]{integer}} expected. Species code identification.}

\item{\code{species_fao_code}}{Object of class {\link[base]{character}} expected. Species code identification on 3 characters.}

\item{\code{sample_standardised_length_class_lf}}{Object of class {\link[base]{integer}} expected. Sample standardised length class length fork of measured individus.}

\item{\code{sample_number_measured_extrapolated_lf}}{Object of class {\link[base]{numeric}} expected. Sample number of measured individus extrapolated to all counted individus.}

\item{\code{sample_total_count}}{Object of class {\link[base]{integer}} expected. Sample number of total individus counted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarysample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarysample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
