% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarysampleraw.R
\name{elementarysampleraw}
\alias{elementarysampleraw}
\title{R6 class trip creation}
\description{
Create R6 reference object class elementarysampleraw
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarysampleraw$new()}}
\item \href{#method-clone}{\code{elementarysampleraw$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarysampleraw class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$new(
  trip_id,
  well_id,
  sample_id,
  sub_sample_id,
  sample_quality,
  sample_type,
  specie_code,
  specie_code3l,
  length_type,
  sample_total_count,
  sample_number_measured,
  sample_length_class
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{well_id}}{(character) Well identification.}

\item{\code{sample_id}}{(character) Sample identification.}

\item{\code{sub_sample_id}}{(integer) Sub sample identification.}

\item{\code{sample_quality}}{(integer) Sample quality identification.}

\item{\code{sample_type}}{(integer) Sample type identification.}

\item{\code{specie_code}}{(integer) Specie code identification.}

\item{\code{specie_code3l}}{(character) Specie code identification on 3 characters.}

\item{\code{length_type}}{(integer) Length type identification, 1 for LD1 and 2 for LF.}

\item{\code{sample_total_count}}{(integer) Sample number of total individus counted.}

\item{\code{sample_number_measured}}{(integer) Sample number of measured individus.}

\item{\code{sample_length_class}}{(integer) Sample length class of measured individus.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
