% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarysampleraw.R
\name{elementarysampleraw}
\alias{elementarysampleraw}
\title{R6 class trip creation}
\description{
Create R6 reference object class elementarysampleraw
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarysampleraw$new()}}
\item \href{#method-clone}{\code{elementarysampleraw$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$new(
  trip_id,
  well_id,
  sample_id,
  sub_sample_id,
  sample_quality,
  sample_type,
  specie_code3l,
  length_type,
  sample_total_count,
  sample_number_measured,
  sample_length_class
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarysampleraw$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
