% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarywellplan.R
\name{elementarywellplan}
\alias{elementarywellplan}
\title{R6 class elementarywellplan}
\description{
Create R6 reference object class elementarywellplan
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elementarywellplan-new}{\code{elementarywellplan$new()}}
\item \href{#method-elementarywellplan-clone}{\code{elementarywellplan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarywellplan-new"></a>}}
\if{latex}{\out{\hypertarget{method-elementarywellplan-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 elementarywellplan class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarywellplan$new(
  wellplan_id,
  well_id,
  activity_id,
  sample_id,
  species_fao_code,
  wellplan_weight,
  wellplan_count,
  weight_category_code,
  weight_category_label
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wellplan_id}}{Object of class {\link[base]{character}} expected. Wellplan identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.}

\item{\code{activity_id}}{Object of class {\link[base]{character}} expected. Activity identification.}

\item{\code{sample_id}}{Object of class {\link[base]{character}} expected. Sample identification.}

\item{\code{species_fao_code}}{Object of class {\link[base]{character}} expected. Specie code identification on 3 characters.}

\item{\code{wellplan_weight}}{(numeric) Weight in tonnes filled in the well plan.}

\item{\code{wellplan_count}}{Object of class {\link[base]{integer}} expected. Well plan number of individus.}

\item{\code{weight_category_code}}{Object of class {\link[base]{character}} expected. Well plan category code identification.}

\item{\code{weight_category_label}}{Object of class {\link[base]{character}} expected. Well plan weight category identification.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elementarywellplan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-elementarywellplan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarywellplan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
