% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementarywellplan.R
\name{elementarywellplan}
\alias{elementarywellplan}
\title{R6 class elementarywellplan creation}
\description{
Create R6 reference object class elementarywellplan
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{elementarywellplan$new()}}
\item \href{#method-clone}{\code{elementarywellplan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarywellplan$new(
  wellplan_id,
  well_id,
  activity_id,
  sample_id,
  specie_code3l,
  wellplan_weight,
  wellplan_number,
  wellplan_weigth_category_code,
  wellplan_weigth_category_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{elementarywellplan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
