% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R
\name{full_trips}
\alias{full_trips}
\title{R6 class full_trips creation}
\description{
Create R6 reference object class full_trips
}
\seealso{
\code{\link{suncalc::getSunlightTimes}}
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{full_trips}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-create_full_trips}{\code{full_trips$create_full_trips()}}
\item \href{#method-filter_by_periode}{\code{full_trips$filter_by_periode()}}
\item \href{#method-add_activities}{\code{full_trips$add_activities()}}
\item \href{#method-add_elementarycatches}{\code{full_trips$add_elementarycatches()}}
\item \href{#method-add_elementarylandings}{\code{full_trips$add_elementarylandings()}}
\item \href{#method-add_wells_samples}{\code{full_trips$add_wells_samples()}}
\item \href{#method-rf1}{\code{full_trips$rf1()}}
\item \href{#method-rf2}{\code{full_trips$rf2()}}
\item \href{#method-conversion_weigth_category}{\code{full_trips$conversion_weigth_category()}}
\item \href{#method-set_count}{\code{full_trips$set_count()}}
\item \href{#method-set_duration}{\code{full_trips$set_duration()}}
\item \href{#method-time_at_sea}{\code{full_trips$time_at_sea()}}
\item \href{#method-fishing_time}{\code{full_trips$fishing_time()}}
\item \href{#method-searching_time}{\code{full_trips$searching_time()}}
\item \href{#method-sample_number_measured_extrapolation}{\code{full_trips$sample_number_measured_extrapolation()}}
\item \href{#method-sample_length_class_ld1_to_lf}{\code{full_trips$sample_length_class_ld1_to_lf()}}
\item \href{#method-sample_length_class_step_standardisation}{\code{full_trips$sample_length_class_step_standardisation()}}
\item \href{#method-well_set_weigth_categories}{\code{full_trips$well_set_weigth_categories()}}
\item \href{#method-standardised_sample_creation}{\code{full_trips$standardised_sample_creation()}}
\item \href{#method-standardised_sample_set_creation}{\code{full_trips$standardised_sample_set_creation()}}
\item \href{#method-raised_factors_determination}{\code{full_trips$raised_factors_determination()}}
\item \href{#method-raised_standardised_sample_set}{\code{full_trips$raised_standardised_sample_set()}}
\item \href{#method-path_to_level3}{\code{full_trips$path_to_level3()}}
\item \href{#method-show_me_what_you_got}{\code{full_trips$show_me_what_you_got()}}
\item \href{#method-clone}{\code{full_trips$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="add">}\href{../../t3/html/list_t3.html#method-add}{\code{t3::list_t3$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count">}\href{../../t3/html/list_t3.html#method-count}{\code{t3::list_t3$count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract">}\href{../../t3/html/list_t3.html#method-extract}{\code{t3::list_t3$extract()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter">}\href{../../t3/html/list_t3.html#method-filter}{\code{t3::list_t3$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="initialize">}\href{../../t3/html/list_t3.html#method-initialize}{\code{t3::list_t3$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name">}\href{../../t3/html/list_t3.html#method-name}{\code{t3::list_t3$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove">}\href{../../t3/html/list_t3.html#method-remove}{\code{t3::list_t3$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view">}\href{../../t3/html/list_t3.html#method-view}{\code{t3::list_t3$view()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_full_trips"></a>}}
\subsection{Method \code{create_full_trips()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$create_full_trips(object_trips)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_by_periode"></a>}}
\subsection{Method \code{filter_by_periode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$filter_by_periode(periode_reference)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_activities"></a>}}
\subsection{Method \code{add_activities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_activities(object_activities)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_elementarycatches"></a>}}
\subsection{Method \code{add_elementarycatches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarycatches(object_elementarycatches)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_elementarylandings"></a>}}
\subsection{Method \code{add_elementarylandings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarylandings(object_elementarylandings)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_wells_samples"></a>}}
\subsection{Method \code{add_wells_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_wells_samples(object_wells)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rf1"></a>}}
\subsection{Method \code{rf1()}}{
Process of Raising Factor level 1 (rf1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf1(species_rf1, rf1_lowest_limit = 0.8, rf1_highest_limit = 1.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species_rf1}}{(character) ISO code on 3 letters related to one or more species.}

\item{\code{rf1_lowest_limit}}{(numeric) Verification value for the lowest limit of the rf1.}

\item{\code{rf1_highest_limit}}{(numeric) Verification value for the highest limit of the rf1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rf2"></a>}}
\subsection{Method \code{rf2()}}{
Process of Raising Factor level 2 (rf2).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conversion_weigth_category"></a>}}
\subsection{Method \code{conversion_weigth_category()}}{
Process of logbook weigth categories conversion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$conversion_weigth_category()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_count"></a>}}
\subsection{Method \code{set_count()}}{
Process for postive sets count.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_duration"></a>}}
\subsection{Method \code{set_duration()}}{
Process for set duration calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_duration(set_duration_ref)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_duration_ref}}{(data frame) Data and parameters for set duration calculation (by year, country, ocean and school type).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time_at_sea"></a>}}
\subsection{Method \code{time_at_sea()}}{
Process for time at sea calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$time_at_sea()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fishing_time"></a>}}
\subsection{Method \code{fishing_time()}}{
Process for fishing time calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$fishing_time(sunrise_schema = "sunrise", sunset_schema = "sunset")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sunrise_schema}}{(character) Sunrise caracteristic. By default "sunrise" (top edge of the sun appears on the horizon). See below for more details.}

\item{\code{sunset_schema}}{(character) Sunset caracteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See below for more details.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Available variables are:
\itemize{
 \item{"sunrise"}{sunrise (top edge of the sun appears on the horizon)}
 \item{"sunriseEnd"}{sunrise ends (bottom edge of the sun touches the horizon)}
 \item{"goldenHourEnd"}{morning golden hour (soft light, best time for photography) ends}
 \item{"solarNoon"}{solar noon (sun is in the highest position)}
 \item{"goldenHour"}{evening golden hour starts}
 \item{"sunsetStart"}{sunset starts (bottom edge of the sun touches the horizon)}
 \item{"sunset"}{sunset (sun disappears below the horizon, evening civil twilight starts)}
 \item{"dusk"}{dusk (evening nautical twilight starts)}
 \item{"nauticalDusk"}{nautical dusk (evening astronomical twilight starts)}
 \item{"night"}{night starts (dark enough for astronomical observations)}
 \item{"nadir"}{nadir (darkest moment of the night, sun is in the lowest position)}
 \item{"nightEnd"}{night ends (morning astronomical twilight starts)}
 \item{"nauticalDawn"}{nautical dawn (morning nautical twilight starts)}
 \item{"dawn"}{dawn (morning nautical twilight ends, morning civil twilight starts)}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-searching_time"></a>}}
\subsection{Method \code{searching_time()}}{
Process for searching time calculation (in hours, fishing time minus sets durations).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$searching_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_number_measured_extrapolation"></a>}}
\subsection{Method \code{sample_number_measured_extrapolation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_number_measured_extrapolation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_length_class_ld1_to_lf"></a>}}
\subsection{Method \code{sample_length_class_ld1_to_lf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_ld1_to_lf(length_step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_length_class_step_standardisation"></a>}}
\subsection{Method \code{sample_length_class_step_standardisation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_step_standardisation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-well_set_weigth_categories"></a>}}
\subsection{Method \code{well_set_weigth_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$well_set_weigth_categories(sample_set)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardised_sample_creation"></a>}}
\subsection{Method \code{standardised_sample_creation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_creation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardised_sample_set_creation"></a>}}
\subsection{Method \code{standardised_sample_set_creation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_set_creation(length_weight_relationship_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raised_factors_determination"></a>}}
\subsection{Method \code{raised_factors_determination()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_factors_determination(
  threshold_rf_minus10,
  threshold_rf_plus10,
  threshold_frequency_rf_minus10,
  threshold_frequency_rf_plus10,
  threshold_rf_total
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raised_standardised_sample_set"></a>}}
\subsection{Method \code{raised_standardised_sample_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_standardised_sample_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path_to_level3"></a>}}
\subsection{Method \code{path_to_level3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$path_to_level3()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_me_what_you_got"></a>}}
\subsection{Method \code{show_me_what_you_got()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$show_me_what_you_got()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
