% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R
\name{full_trips}
\alias{full_trips}
\title{R6 class full_trips creation}
\description{
Create R6 reference object class full_trips
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{full_trips}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-create_full_trips}{\code{full_trips$create_full_trips()}}
\item \href{#method-filter_by_periode}{\code{full_trips$filter_by_periode()}}
\item \href{#method-add_activities}{\code{full_trips$add_activities()}}
\item \href{#method-add_elementarycatches}{\code{full_trips$add_elementarycatches()}}
\item \href{#method-add_elementarylandings}{\code{full_trips$add_elementarylandings()}}
\item \href{#method-add_samples}{\code{full_trips$add_samples()}}
\item \href{#method-rf1}{\code{full_trips$rf1()}}
\item \href{#method-rf2}{\code{full_trips$rf2()}}
\item \href{#method-conversion_weigth_category}{\code{full_trips$conversion_weigth_category()}}
\item \href{#method-set_count}{\code{full_trips$set_count()}}
\item \href{#method-set_duration}{\code{full_trips$set_duration()}}
\item \href{#method-time_at_sea}{\code{full_trips$time_at_sea()}}
\item \href{#method-fishing_time}{\code{full_trips$fishing_time()}}
\item \href{#method-searching_time}{\code{full_trips$searching_time()}}
\item \href{#method-sample_number_measured_extrapolation}{\code{full_trips$sample_number_measured_extrapolation()}}
\item \href{#method-sample_length_class_ld1_to_lf}{\code{full_trips$sample_length_class_ld1_to_lf()}}
\item \href{#method-sample_length_class_step_standardisation}{\code{full_trips$sample_length_class_step_standardisation()}}
\item \href{#method-clone}{\code{full_trips$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="add">}\href{../../t3/html/list_t3.html#method-add}{\code{t3::list_t3$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count">}\href{../../t3/html/list_t3.html#method-count}{\code{t3::list_t3$count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract">}\href{../../t3/html/list_t3.html#method-extract}{\code{t3::list_t3$extract()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter">}\href{../../t3/html/list_t3.html#method-filter}{\code{t3::list_t3$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="initialize">}\href{../../t3/html/list_t3.html#method-initialize}{\code{t3::list_t3$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name">}\href{../../t3/html/list_t3.html#method-name}{\code{t3::list_t3$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove">}\href{../../t3/html/list_t3.html#method-remove}{\code{t3::list_t3$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view">}\href{../../t3/html/list_t3.html#method-view}{\code{t3::list_t3$view()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_full_trips"></a>}}
\subsection{Method \code{create_full_trips()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$create_full_trips(object_trips)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_by_periode"></a>}}
\subsection{Method \code{filter_by_periode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$filter_by_periode(periode_reference)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_activities"></a>}}
\subsection{Method \code{add_activities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_activities(object_activities)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_elementarycatches"></a>}}
\subsection{Method \code{add_elementarycatches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarycatches(object_elementarycatches)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_elementarylandings"></a>}}
\subsection{Method \code{add_elementarylandings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarylandings(object_elementarylandings)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_samples"></a>}}
\subsection{Method \code{add_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_samples(object_samples)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rf1"></a>}}
\subsection{Method \code{rf1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf1(species_rf1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rf2"></a>}}
\subsection{Method \code{rf2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conversion_weigth_category"></a>}}
\subsection{Method \code{conversion_weigth_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$conversion_weigth_category()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_count"></a>}}
\subsection{Method \code{set_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_duration"></a>}}
\subsection{Method \code{set_duration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_duration(set_duration_ref)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time_at_sea"></a>}}
\subsection{Method \code{time_at_sea()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$time_at_sea()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fishing_time"></a>}}
\subsection{Method \code{fishing_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$fishing_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-searching_time"></a>}}
\subsection{Method \code{searching_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$searching_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_number_measured_extrapolation"></a>}}
\subsection{Method \code{sample_number_measured_extrapolation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_number_measured_extrapolation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_length_class_ld1_to_lf"></a>}}
\subsection{Method \code{sample_length_class_ld1_to_lf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_ld1_to_lf(length_step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_length_class_step_standardisation"></a>}}
\subsection{Method \code{sample_length_class_step_standardisation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_step_standardisation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
