% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R
\name{full_trips}
\alias{full_trips}
\title{R6 class full_trips}
\description{
Create R6 reference object class full_trips
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{full_trips}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-full_trips-create_full_trips}{\code{full_trips$create_full_trips()}}
\item \href{#method-full_trips-filter_by_periode}{\code{full_trips$filter_by_periode()}}
\item \href{#method-full_trips-add_activities}{\code{full_trips$add_activities()}}
\item \href{#method-full_trips-add_elementarycatches}{\code{full_trips$add_elementarycatches()}}
\item \href{#method-full_trips-add_elementarylandings}{\code{full_trips$add_elementarylandings()}}
\item \href{#method-full_trips-add_wells_samples}{\code{full_trips$add_wells_samples()}}
\item \href{#method-full_trips-rf1}{\code{full_trips$rf1()}}
\item \href{#method-full_trips-rf2}{\code{full_trips$rf2()}}
\item \href{#method-full_trips-conversion_weigth_category}{\code{full_trips$conversion_weigth_category()}}
\item \href{#method-full_trips-set_count}{\code{full_trips$set_count()}}
\item \href{#method-full_trips-set_duration}{\code{full_trips$set_duration()}}
\item \href{#method-full_trips-time_at_sea}{\code{full_trips$time_at_sea()}}
\item \href{#method-full_trips-fishing_time}{\code{full_trips$fishing_time()}}
\item \href{#method-full_trips-searching_time}{\code{full_trips$searching_time()}}
\item \href{#method-full_trips-sample_length_class_ld1_to_lf}{\code{full_trips$sample_length_class_ld1_to_lf()}}
\item \href{#method-full_trips-sample_number_measured_extrapolation}{\code{full_trips$sample_number_measured_extrapolation()}}
\item \href{#method-full_trips-sample_length_class_step_standardisation}{\code{full_trips$sample_length_class_step_standardisation()}}
\item \href{#method-full_trips-well_set_weigth_categories}{\code{full_trips$well_set_weigth_categories()}}
\item \href{#method-full_trips-standardised_sample_creation}{\code{full_trips$standardised_sample_creation()}}
\item \href{#method-full_trips-standardised_sample_set_creation}{\code{full_trips$standardised_sample_set_creation()}}
\item \href{#method-full_trips-raised_factors_determination}{\code{full_trips$raised_factors_determination()}}
\item \href{#method-full_trips-raised_standardised_sample_set}{\code{full_trips$raised_standardised_sample_set()}}
\item \href{#method-full_trips-path_to_level3}{\code{full_trips$path_to_level3()}}
\item \href{#method-full_trips-data_preparatory}{\code{full_trips$data_preparatory()}}
\item \href{#method-full_trips-random_forest_models}{\code{full_trips$random_forest_models()}}
\item \href{#method-full_trips-models_checking}{\code{full_trips$models_checking()}}
\item \href{#method-full_trips-data_formatting_for_predictions}{\code{full_trips$data_formatting_for_predictions()}}
\item \href{#method-full_trips-model_predictions}{\code{full_trips$model_predictions()}}
\item \href{#method-full_trips-show_me_what_you_got}{\code{full_trips$show_me_what_you_got()}}
\item \href{#method-full_trips-clone}{\code{full_trips$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="add"><a href='../../t3/html/list_t3.html#method-list_t3-add'><code>t3::list_t3$add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count"><a href='../../t3/html/list_t3.html#method-list_t3-count'><code>t3::list_t3$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract"><a href='../../t3/html/list_t3.html#method-list_t3-extract'><code>t3::list_t3$extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract_l1_element_value"><a href='../../t3/html/list_t3.html#method-list_t3-extract_l1_element_value'><code>t3::list_t3$extract_l1_element_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter"><a href='../../t3/html/list_t3.html#method-list_t3-filter'><code>t3::list_t3$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter_l1"><a href='../../t3/html/list_t3.html#method-list_t3-filter_l1'><code>t3::list_t3$filter_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="initialize"><a href='../../t3/html/list_t3.html#method-list_t3-initialize'><code>t3::list_t3$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="modification_l1"><a href='../../t3/html/list_t3.html#method-list_t3-modification_l1'><code>t3::list_t3$modification_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name"><a href='../../t3/html/list_t3.html#method-list_t3-name'><code>t3::list_t3$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove"><a href='../../t3/html/list_t3.html#method-list_t3-remove'><code>t3::list_t3$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view"><a href='../../t3/html/list_t3.html#method-list_t3-view'><code>t3::list_t3$view()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-create_full_trips"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-create_full_trips}{}}}
\subsection{Method \code{create_full_trips()}}{
Creation of full trip item from trips.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$create_full_trips(object_trips)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_trips}}{Object of type R6-trips expected. A R6 reference object of class trips.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-filter_by_periode"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-filter_by_periode}{}}}
\subsection{Method \code{filter_by_periode()}}{
Function for filter full trips by a reference periode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$filter_by_periode(periode_reference)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{periode_reference}}{Object of class {\link[base]{integer}} expected. Year(s) in 4 digits format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_activities"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_activities}{}}}
\subsection{Method \code{add_activities()}}{
Function for add activities in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_activities(object_activities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_activities}}{Object of type R6-activities expected. A R6 reference object of class activities.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_elementarycatches"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_elementarycatches}{}}}
\subsection{Method \code{add_elementarycatches()}}{
Function for add elementary catches in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarycatches(object_elementarycatches)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_elementarycatches}}{Object of type R6-elementarycatches expected. A R6 reference object of class elementarycatches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_elementarylandings"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_elementarylandings}{}}}
\subsection{Method \code{add_elementarylandings()}}{
Function for add elementary landings in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarylandings(object_elementarylandings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_elementarylandings}}{Object of type R6-elementarylandings expected. A R6 reference object of class elementarylandings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_wells_samples"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_wells_samples}{}}}
\subsection{Method \code{add_wells_samples()}}{
Function for add wells and samples caracteristics in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_wells_samples(object_wells)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_wells}}{Object of type R6-wells expected. A R6 reference object of class wells.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-rf1"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-rf1}{}}}
\subsection{Method \code{rf1()}}{
Process of Raising Factor level 1 (RF1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf1(
  species_rf1 = as.integer(c(1, 2, 3, 4, 9, 11)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species_rf1}}{Object of type \code{\link[base]{integer}} expected. Specie(s) code(s) used for the RF1 process. By default 1 (YFT), 2 (SKJ), 3 (BET), 4 (ALB), 9 (MIX) and 11 (LOT).}

\item{\code{rf1_lowest_limit}}{Object of type \code{\link[base]{numeric}} expected. Verification value for the lowest limit of the RF1. By default 0.8.}

\item{\code{rf1_highest_limit}}{Object of type \code{\link[base]{numeric}} expected. Verification value for the highest limit of the RF1. By default 1.2.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-rf2"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-rf2}{}}}
\subsection{Method \code{rf2()}}{
Process of Raising Factor level 2 (rf2).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf2(global_outputs_path = NULL, outputs_format = "eu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-conversion_weigth_category"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-conversion_weigth_category}{}}}
\subsection{Method \code{conversion_weigth_category()}}{
Process of logbook weigth categories conversion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$conversion_weigth_category(
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-set_count"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-set_count}{}}}
\subsection{Method \code{set_count()}}{
Process for postive sets count.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_count(global_outputs_path = NULL, outputs_format = "eu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-set_duration"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-set_duration}{}}}
\subsection{Method \code{set_duration()}}{
Process for set duration calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_duration(
  set_duration_ref,
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_duration_ref}}{Object of type \code{\link[base]{data.frame}} expected. Data and parameters for set duration calculation (by year, country, ocean and school type).}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-time_at_sea"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-time_at_sea}{}}}
\subsection{Method \code{time_at_sea()}}{
Process for time at sea calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$time_at_sea(global_outputs_path = NULL, outputs_format = "eu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-fishing_time"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-fishing_time}{}}}
\subsection{Method \code{fishing_time()}}{
Process for fishing time calculation (in hours).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$fishing_time(
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sunrise_schema}}{Object of class {\link[base]{character}} expected. Sunrise caracteristic. By default "sunrise" (top edge of the sun appears on the horizon). See below for more details.}

\item{\code{sunset_schema}}{Object of class {\link[base]{character}} expected. Sunset caracteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See below for more details.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Available variables are:
\itemize{
 \item{"sunrise"}{sunrise (top edge of the sun appears on the horizon)}
 \item{"sunriseEnd"}{sunrise ends (bottom edge of the sun touches the horizon)}
 \item{"goldenHourEnd"}{morning golden hour ends(soft light, best time for photography)}
 \item{"solarNoon"}{solar noon (sun is in the highest position)}
 \item{"goldenHour"}{evening golden hour starts}
 \item{"sunsetStart"}{sunset starts (bottom edge of the sun touches the horizon)}
 \item{"sunset"}{sunset (sun disappears below the horizon, evening civil twilight starts)}
 \item{"dusk"}{dusk (evening nautical twilight starts)}
 \item{"nauticalDusk"}{nautical dusk (evening astronomical twilight starts)}
 \item{"night"}{night starts (dark enough for astronomical observations)}
 \item{"nadir"}{nadir (darkest moment of the night, sun is in the lowest position)}
 \item{"nightEnd"}{night ends (morning astronomical twilight starts)}
 \item{"nauticalDawn"}{nautical dawn (morning nautical twilight starts)}
 \item{"dawn"}{dawn (morning nautical twilight ends, morning civil twilight starts)}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-searching_time"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-searching_time}{}}}
\subsection{Method \code{searching_time()}}{
Process for searching time calculation (in hours, fishing time minus sets durations).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$searching_time(global_outputs_path = NULL, outputs_format = "eu")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_length_class_ld1_to_lf"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_length_class_ld1_to_lf}{}}}
\subsection{Method \code{sample_length_class_ld1_to_lf()}}{
Process for length conversion, if necessary, in length fork (lf). Furthermore, variable "sample_number_measured_extrapolated" of process 2.1 will converse in variable "sample_number_measured_extrapolated_lf" (Notably due to the creation of new lf classes during some conversions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_ld1_to_lf(
  length_step,
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length_step}}{Object of type \code{\link[base]{data.frame}} expected. Data frame object with length ratio between ld1 and lf class.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_number_measured_extrapolation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_number_measured_extrapolation}{}}}
\subsection{Method \code{sample_number_measured_extrapolation()}}{
Process for sample number measured individuals extrapolation to sample number individuals counted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_number_measured_extrapolation(
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_length_class_step_standardisation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_length_class_step_standardisation}{}}}
\subsection{Method \code{sample_length_class_step_standardisation()}}{
Process for step standardisation of lf length class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_step_standardisation(
  maximum_lf_class = as.integer(500),
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maximum_lf_class}}{Object of type \code{\link[base]{integer}} expected. Theorical maximum lf class that can occur (all species considerated). By default 500.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-well_set_weigth_categories"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-well_set_weigth_categories}{}}}
\subsection{Method \code{well_set_weigth_categories()}}{
Process for well set weigth categories definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$well_set_weigth_categories(
  sample_set,
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_set}}{Object of type \code{\link[base]{data.frame}} expected. Data frame object with weighted weigh of each set sampled.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-standardised_sample_creation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-standardised_sample_creation}{}}}
\subsection{Method \code{standardised_sample_creation()}}{
Object standardised sample creation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_creation(
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-standardised_sample_set_creation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-standardised_sample_set_creation}{}}}
\subsection{Method \code{standardised_sample_set_creation()}}{
R6 object standardised sample set creation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_set_creation(
  length_weight_relationship_data,
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length_weight_relationship_data}}{Object of type \code{\link[base]{data.frame}} expected. Data frame object with parameters for length weight relationship.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-raised_factors_determination"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-raised_factors_determination}{}}}
\subsection{Method \code{raised_factors_determination()}}{
Raised factors determination for weigth sample set to set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_factors_determination(
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250),
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold_rf_minus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limite value for raising factor on individuals category minus 10. By default 500.}

\item{\code{threshold_rf_plus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limite value for raising factor on individuals category plus 10. By default 500.}

\item{\code{threshold_frequency_rf_minus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limite frequency value for raising factor on individuals category minus 10. By default 75.}

\item{\code{threshold_frequency_rf_plus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limite frequency value for raising factor on individuals category plus 10. By default 75.}

\item{\code{threshold_rf_total}}{Object of type \code{\link[base]{integer}} expected. Threshold limite value for raising factor (all categories). By default 250.}

\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-raised_standardised_sample_set"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-raised_standardised_sample_set}{}}}
\subsection{Method \code{raised_standardised_sample_set()}}{
Application of process 2.8 raised factors on standardised sample set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_standardised_sample_set(
  global_outputs_path = NULL,
  outputs_format = "eu"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_outputs_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction egual TRUE. Path of the global outputs directory. The function will create subsection if necessary.}

\item{\code{outputs_format}}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-path_to_level3"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-path_to_level3}{}}}
\subsection{Method \code{path_to_level3()}}{
Temporary link to the R object model with Antoine D. modelisation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$path_to_level3()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-data_preparatory"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-data_preparatory}{}}}
\subsection{Method \code{data_preparatory()}}{
Data preparatory for the t3 modelling process (level 3).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$data_preparatory(
  inputs_level3 = NULL,
  inputs_level3_path = NULL,
  outputs_directory,
  periode_reference = NULL,
  target_year = as.integer(lubridate::year(Sys.time() - 1)),
  period_duration = 4L,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs_level3}}{Object of type \code{\link[base]{data.frame}} expected. Inputs of levels 3 (see function path to level 3).}

\item{\code{inputs_level3_path}}{Object of type \code{\link[base]{character}} expected. Path to the folder containing yearly data output of the level 1 and 2 (output of the function the path to level 3). If provide, replace the inputs_level3 object.}

\item{\code{outputs_directory}}{Object of type \code{\link[base]{character}} expected. Path of the t3 processes outputs directory.}

\item{\code{periode_reference}}{Object of type \code{\link[base]{integer}} expected. Year(s) period of reference for modelling estimation.}

\item{\code{target_year}}{Object of type \code{\link[base]{integer}} expected. Year of interest for the model estimation and prediction.Default value is current year -1.}

\item{\code{period_duration}}{Object of type \code{\link[base]{integer}} expected. number of years use for the modelling. The default value is 5}

\item{\code{distance_maximum}}{Object of type \code{\link[base]{integer}} expected. Maximum distance between all sets of a sampled well. By default 5.}

\item{\code{number_sets_maximum}}{Object of type \code{\link[base]{integer}} expected. Maximum number of sets allowed in mixture. By default 5.}

\item{\code{set_weight_minimum}}{Object of type \code{\link[base]{integer}} expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.}

\item{\code{minimum_set_frequency}}{Object of type \code{\link[base]{numeric}} expected. Minimum threshold proportion of set in awell to be used for model training in the process. By default 0.1.}

\item{\code{vessel_id_ignored}}{Object of type \code{\link[base]{integer}} expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-random_forest_models"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-random_forest_models}{}}}
\subsection{Method \code{random_forest_models()}}{
Modelling proportions in sets througth random forest models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$random_forest_models(
  outputs_level3_process1,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs_level3_process1}}{Object of type \code{\link[base]{data.frame}} expected. Output table data_lb_sample_screened from process 3.1.}

\item{\code{num.trees}}{Object of type \code{\link[base]{integer}} expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.}

\item{\code{mtry}}{Object of type \code{\link[base]{integer}} expected. Number of variables randomly sampled as candidates at each split. The default value is 2.}

\item{\code{min.node.size}}{Object of type \code{\link[base]{numeric}} expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.}

\item{\code{seed_number}}{Object of type \code{\link[base]{integer}} expected. Set the initial seed for the modelling. The default value is 7.}

\item{\code{small_fish_only}}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-models_checking"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-models_checking}{}}}
\subsection{Method \code{models_checking()}}{
Load each full model and compute figures and tables to check the model quality. Furthermore, create a map of samples used for each model and relationship between logbook reports and samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$models_checking(
  outputs_level3_process2,
  outputs_directory,
  plot_sample = F,
  avdth_patch_coord = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs_level3_process2}}{Object of type \code{\link[base]{list}} expected. Outputs models and data from process 3.2.}

\item{\code{outputs_directory}}{Object of type \code{\link[base]{character}} expected. Outputs directory path.}

\item{\code{plot_sample}}{\code{\link[base]{logical}}. Whether the sample figure is computed. Default value = F}

\item{\code{avdth_patch_coord}}{parameter waiting for coordinate conversion patch from avdth database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-data_formatting_for_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-data_formatting_for_predictions}{}}}
\subsection{Method \code{data_formatting_for_predictions()}}{
Formatting data for model predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$data_formatting_for_predictions(
  inputs_level3,
  outputs_level3_process1,
  target_year,
  vessel_id_ignored = NULL,
  small_fish_only = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs_level3}}{Object of type \code{\link[base]{data.frame}} expected. Inputs of levels 3 (see function path to level 3).}

\item{\code{outputs_level3_process1}}{Object of type \code{\link[base]{data.frame}} expected. Output table data_lb_sample_screened from process 3.1.}

\item{\code{target_year}}{Object of type \code{\link[base]{integer}} expected. The year of interest for the model estimation and prediction.}

\item{\code{vessel_id_ignored}}{Object of type \code{\link[base]{integer}} expected. Specify here vessel(s) id(s) if you want to ignore it in the model estimation and prediction .By default NULL.}

\item{\code{small_fish_only}}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-model_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-model_predictions}{}}}
\subsection{Method \code{model_predictions()}}{
Model predictions for the species composition and computing of catches.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$model_predictions(
  outputs_level3_process2,
  outputs_level3_process4,
  outputs_directory,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outputs_level3_process2}}{Object of type \code{\link[base]{list}} expected. Outputs from level 3 process 2 (random forest models).}

\item{\code{outputs_level3_process4}}{Object of type \code{\link[base]{list}} expected. Outputs from level 3 process 4 (data formatting for predictions).}

\item{\code{outputs_directory}}{Object of type \code{\link[base]{character}} expected. Outputs directory path.}

\item{\code{ci}}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.}

\item{\code{ci_type}}{Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if  the ci parameter is TRUE.}

\item{\code{Nboot}}{Object of type \code{\link[base]{numeric}} expected. The number of bootstrap samples desired for the ci computation. The default value is 10.}

\item{\code{plot_predict}}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether maps of catch at size have to be done.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-show_me_what_you_got"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-show_me_what_you_got}{}}}
\subsection{Method \code{show_me_what_you_got()}}{
Most powerfull and "schwifty" function in the univers for "open the T3 process" and manipulate in live R6 objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$show_me_what_you_got()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-clone"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
