% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R
\name{full_trips}
\alias{full_trips}
\title{R6 class full_trips}
\description{
Create R6 reference object class full_trips
}
\section{Super class}{
\code{\link[t3:list_t3]{t3::list_t3}} -> \code{full_trips}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-full_trips-create_full_trips}{\code{full_trips$create_full_trips()}}
\item \href{#method-full_trips-add_activities}{\code{full_trips$add_activities()}}
\item \href{#method-full_trips-filter_by_years_period}{\code{full_trips$filter_by_years_period()}}
\item \href{#method-full_trips-add_elementarylandings}{\code{full_trips$add_elementarylandings()}}
\item \href{#method-full_trips-add_wells_samples}{\code{full_trips$add_wells_samples()}}
\item \href{#method-full_trips-rf1}{\code{full_trips$rf1()}}
\item \href{#method-full_trips-conversion_weight_category}{\code{full_trips$conversion_weight_category()}}
\item \href{#method-full_trips-set_count}{\code{full_trips$set_count()}}
\item \href{#method-full_trips-fishing_effort}{\code{full_trips$fishing_effort()}}
\item \href{#method-full_trips-sample_length_class_ld1_to_lf}{\code{full_trips$sample_length_class_ld1_to_lf()}}
\item \href{#method-full_trips-sample_number_measured_extrapolation}{\code{full_trips$sample_number_measured_extrapolation()}}
\item \href{#method-full_trips-sample_length_class_step_standardisation}{\code{full_trips$sample_length_class_step_standardisation()}}
\item \href{#method-full_trips-well_set_weight_categories}{\code{full_trips$well_set_weight_categories()}}
\item \href{#method-full_trips-standardised_sample_creation}{\code{full_trips$standardised_sample_creation()}}
\item \href{#method-full_trips-standardised_sample_set_creation}{\code{full_trips$standardised_sample_set_creation()}}
\item \href{#method-full_trips-raised_factors_determination}{\code{full_trips$raised_factors_determination()}}
\item \href{#method-full_trips-raised_standardised_sample_set}{\code{full_trips$raised_standardised_sample_set()}}
\item \href{#method-full_trips-path_to_level3}{\code{full_trips$path_to_level3()}}
\item \href{#method-full_trips-data_preparatory}{\code{full_trips$data_preparatory()}}
\item \href{#method-full_trips-random_forest_models}{\code{full_trips$random_forest_models()}}
\item \href{#method-full_trips-models_checking}{\code{full_trips$models_checking()}}
\item \href{#method-full_trips-data_formatting_for_predictions}{\code{full_trips$data_formatting_for_predictions()}}
\item \href{#method-full_trips-model_predictions}{\code{full_trips$model_predictions()}}
\item \href{#method-full_trips-show_me_what_you_got}{\code{full_trips$show_me_what_you_got()}}
\item \href{#method-full_trips-clone}{\code{full_trips$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="add"><a href='../../t3/html/list_t3.html#method-list_t3-add'><code>t3::list_t3$add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="count"><a href='../../t3/html/list_t3.html#method-list_t3-count'><code>t3::list_t3$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract"><a href='../../t3/html/list_t3.html#method-list_t3-extract'><code>t3::list_t3$extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="extract_l1_element_value"><a href='../../t3/html/list_t3.html#method-list_t3-extract_l1_element_value'><code>t3::list_t3$extract_l1_element_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter"><a href='../../t3/html/list_t3.html#method-list_t3-filter'><code>t3::list_t3$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="filter_l1"><a href='../../t3/html/list_t3.html#method-list_t3-filter_l1'><code>t3::list_t3$filter_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="initialize"><a href='../../t3/html/list_t3.html#method-list_t3-initialize'><code>t3::list_t3$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="modification_l1"><a href='../../t3/html/list_t3.html#method-list_t3-modification_l1'><code>t3::list_t3$modification_l1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="name"><a href='../../t3/html/list_t3.html#method-list_t3-name'><code>t3::list_t3$name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="remove"><a href='../../t3/html/list_t3.html#method-list_t3-remove'><code>t3::list_t3$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="t3" data-topic="list_t3" data-id="view"><a href='../../t3/html/list_t3.html#method-list_t3-view'><code>t3::list_t3$view()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-create_full_trips"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-create_full_trips}{}}}
\subsection{Method \code{create_full_trips()}}{
Creation of full trip item from trips.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$create_full_trips(object_trips)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_trips}}{Object of type R6-trips expected. A R6 reference object of class trips.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_activities"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_activities}{}}}
\subsection{Method \code{add_activities()}}{
Function for add activities in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_activities(object_activities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_activities}}{Object of type R6-activities expected. A R6 reference object of class activities.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-filter_by_years_period"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-filter_by_years_period}{}}}
\subsection{Method \code{filter_by_years_period()}}{
Function for filter full trips by a year(s) period.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$filter_by_years_period(years_period)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{years_period}}{Object of class {\link[base]{integer}} expected. Year(s) in 4 digits format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_elementarylandings"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_elementarylandings}{}}}
\subsection{Method \code{add_elementarylandings()}}{
Function for add elementary landings in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_elementarylandings(object_elementarylandings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_elementarylandings}}{Object of type R6-elementarylandings expected. A R6 reference object of class elementarylandings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-add_wells_samples"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-add_wells_samples}{}}}
\subsection{Method \code{add_wells_samples()}}{
Function for add wells and samples caracteristics in full trips object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$add_wells_samples(object_wells)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_wells}}{Object of type R6-wells expected. A R6 reference object of class wells.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-rf1"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-rf1}{}}}
\subsection{Method \code{rf1()}}{
Process of Raising Factor level 1 (RF1 and RF2) calculation.
The RF1 aim to correct the weight visual estimation bias of catches filled in logbooks and correspond to the ratio of landing weight on catch weight, of the species defined by the \code{species_fao_codes_rf1} argument.
Raising Factor 2 (RF2) is not implemented yet but will be computed at this step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$rf1(
  rf1_computation = TRUE,
  apply_rf1_on_bycatch = TRUE,
  species_fao_codes_rf1 = c("YFT", "SKJ", "BET", "ALB", "LOT", "MIX", "TUN"),
  species_fate_codes_rf1 = as.integer(6),
  vessel_type_codes_rf1 = as.integer(c(4, 5, 6)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  global_output_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rf1_computation}}{Object of class \code{\link[base]{logical}} expected. If FALSE rf1 is not calculated (rf1=1 for all trips).
By default TRUE, the rf1 is calculated for each trip.}

\item{\code{apply_rf1_on_bycatch}}{Object of class \code{\link[base]{logical}} expected. By default TRUE, rf1 values will be applied to all the logbook catches associated to the trip, including by-catch species.
If FALSE, only the catch weights of species belonging to the species list, defined by the \code{species_fao_codes_rf1} argument are corrected, rf1 is not applied to by-catch species.}

\item{\code{species_fao_codes_rf1}}{Object of type \code{\link[base]{character}} expected.Specie(s) FAO code(s) used for the RF1 process.
By default, use codes YFT (\emph{Thunnus albacares}), SKJ (\emph{Katsuwonus pelamis}), BET (\emph{Thunnus obesus}), ALB (\emph{Thunnus alalunga}),
LOT (\emph{Thunnus tonggol}) and TUN/MIX (mix of tunas species in Observe/AVDTH database) (French and Mayotte fleets).}

\item{\code{species_fate_codes_rf1}}{Object of type \code{\link[base]{integer}} expected. By default 6 ("Retained, presumably destined for the cannery"). Specie(s) fate code(s) used for the RF1 process.}

\item{\code{vessel_type_codes_rf1}}{Object of type \code{\link[base]{integer}} expected. By default 4, 5 and 6. Vessel type(s).}

\item{\code{rf1_lowest_limit}}{Object of type \code{\link[base]{numeric}} expected. Verification value for the lowest limit of the RF1. By default 0.8.}

\item{\code{rf1_highest_limit}}{Object of type \code{\link[base]{numeric}} expected. Verification value for the highest limit of the RF1. By default 1.2.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following outputs are extracted and saved in ".csv" format under the path: "global_output_path/level1/data/". \cr
 \itemize{
 \item{process_1_1_detail: a table (.csv) with as many rows as elementary catches and 23 columns:}
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{activity_id: } activity identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{activity_latitude: } activity latitude, type \code{\link[base]{numeric}}.
 \item{activity_longitude: } activity longitude, type \code{\link[base]{numeric}}.
 \item{trip_end_date: } trip end date (y-m-d format), type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{rf1: } raising factor to correct the weight visual estimation bias of catches filled in logbooks.
 Rf1 is the ratio of landing weight on catch weight, of the species defined by the \code{species_fao_codes_rf1} argument.
 , type \code{\link[base]{numeric}}.
 \item{statut_rf1: } status rf1, type \code{\link[base]{numeric}}.
 \item{rf2: } raising factor to correct missing logbook(s) not implemented yet (rf2=1), type \code{\link[base]{numeric}}.
 \item{statut_rf2: } status rf2, type \code{\link[base]{numeric}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{elementarycatch_id: } elementary catch identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{species_fate_code: } species fate codes, type \code{\link[base]{integer}}.
 For example in Observe database :  \itemize{
 \item{4 : } discarded alive.
 \item{5 : } discarded dead.
 \item{6 : } Retained, presumably destined for the cannery
 \item{8 : } used for crew consumption on board.
 \item{11 : } discarded status unknown (only for EMS and logbook).
 \item{15 : } retained for local market or dried/salted fish on board.
 }
 \item{landing_weight: } landing weight (without local market), in tonnes, type \code{\link[base]{numeric}}.
 \item{catch_weight: } catch weight (visual estimation), in tonnes, type \code{\link[base]{numeric}}.
 \item{catch_count: } catch count, type \code{\link[base]{integer}}.
 \item{catch_weight_rf2: } catch weight after visual estimation correction, in tonnes: \code{catch_weight_rf2=catch_weight x rf1 (x rf2)}
 (\href{https://ob7-ird.github.io/t3/articles/level_1.html#process-1-1-raising-factors-level-1}{Process 1.1: Raising Factors level 1}), type \code{\link[base]{numeric}}.
 \item{statut_rf1_label: } status rf1 label, type \code{\link[base]{character}}.
 \item{statut_rf2_label: } status rf2 label, type \code{\link[base]{character}}.
 }
 \item{process_1_1_global: a table (.csv) with as many rows as full trips and 17 columns:}
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date (y-m-d format), type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{rf1: } raising factor to correct the weight visual estimation bias of catches filled in logbooks.
 Rf1 is the ratio of landing weight on catch weight, of the species defined by the \code{species_fao_codes_rf1} argument.
, type \code{\link[base]{numeric}}.
 \item{statut_rf1: } status rf1, type \code{\link[base]{numeric}}.
 \item{rf2: } raising factor to correct missing logbook(s) not implemented yet (rf2=1), type \code{\link[base]{numeric}}.
 \item{statut_rf2: } status rf2, type \code{\link[base]{numeric}}.
 \item{landing_weight: } landing weight (without local market), in tonnes, type \code{\link[base]{numeric}}.
 \item{catch_weight: } catch weight (visual estimation), in tonnes, type \code{\link[base]{numeric}}.
 \item{catch_count: } catch count, type \code{\link[base]{integer}}.
 \item{catch_weight_rf2: } catch weight after visual estimation correction (tonnes): \code{catch_weight_rf2=catch_weight x rf1 (x rf2)}
 (\href{https://ob7-ird.github.io/t3/reference/full_trips.html#method-full_trips-rf1}{\code{full_trips$rf1()}}), type \code{\link[base]{numeric}}.
 \item{statut_rf1_label: } status rf1 label, type \code{\link[base]{character}}.
 \item{statut_rf2_label: } status rf2 label, type \code{\link[base]{character}}.
 }
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-conversion_weight_category"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-conversion_weight_category}{}}}
\subsection{Method \code{conversion_weight_category()}}{
Process of logbook weight categories conversion. \cr
Logbook's weight categories change from one tuna fishing company to another, that involves overlaps and are hardly usable directly from the logbook.
This process aims to homogenize these weight categories and create simplify categories divided in function of the fishing school and the ocean:
\itemize{
\item{< 10kg and > 10kg} for the floating object school in the Atlantic and Indian Ocean,
\item{ < 10kg, 10-30kg and > 30kg} for undetermined and free school in the Atlantic Ocean,
\item{< 10kg and > 10kg} for undetermined and free school in the Indian Ocean.
}
For each layer ocean/fishing school/specie/logbook weight category, a distribution key is applied for conversion to standardized weight categories.
 Details of the distribution key is available in the vignette \href{https://ob7-ird.github.io/t3/articles/level_1.html#process-1-2-logbook-weight-categories-conversion}{Process 1.2: logbook weight categories conversion}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$conversion_weight_category(
  global_output_path = NULL,
  referential_template = "observe"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}

\item{\code{referential_template}}{Object of class \code{\link[base]{character}} expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level1/data/". \cr
 process_1_2: a table (.csv) with as many rows as elementary catches, plus the catches resulting from the conversion of weight categories and 23 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{activity_id: } activity identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{activity_latitude: } activity latitude, type \code{\link[base]{numeric}}.
 \item{activity_longitude: } activity longitude, type \code{\link[base]{numeric}}.
 \item{activity_date: } activity date, type \code{\link[base]{character}}.
 \item{ocean_code: } ocean code, type \code{\link[base]{integer}}.
  For example \code{ocean_code=1} for the Atlantic Ocean and \code{ocean_code=2} the Indian Ocean.
 \item{school_type_code:} school type code, type \code{\link[base]{integer}}.
  In Observe referential template: 1 for floating object school, 2 for free school and 0 for undetermined school.
 \item{elementarycatch_id: } elementary catch identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{weight_category_code: } weight category code defined in logbooks, type \code{\link[base]{character}}.
 \item{weight_category_min: } weight category's lower limit (kg), type \code{\link[base]{numeric}}.
 \item{weight_category_max: } weight category's upper limit (kg), type \code{\link[base]{numeric}}.
 \item{weight_category_label: } weight category label defined in logbooks, type \code{\link[base]{character}}.
 \item{catch_weight_rf2: } catch weight after visual estimation correction  (tonnes): \code{catch_weight_rf2=catch_weight x rf1 (x rf2)}
 (\href{https://ob7-ird.github.io/t3/articles/level_1.html#process-1-1-raising-factors-level-1}{Process 1.1: Raising Factors level 1}), type \code{\link[base]{numeric}}.
 \item{weight_category_code_corrected: } weight category code after conversion, type \code{\link[base]{character}}.
 \item{catch_weigh_category_code_corrected: } catch weight after weight category conversion (tonnes), type \code{\link[base]{numeric}}.\cr
  In fact, the catch weight corresponding to the logbook weight category can be divided between several corrected weight categories according to the distribution key applied for conversion to standardized weight categories.
 \item{catch_count: } catch count, type \code{\link[base]{integer}}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-set_count"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-set_count}{}}}
\subsection{Method \code{set_count()}}{
Process for positive sets count.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$set_count(
  global_output_path = NULL,
  referential_template = "observe"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}

\item{\code{referential_template}}{Object of class \code{\link[base]{character}} expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level1/data/". \cr
 process_1_3: a table (.csv) with as many rows as activities and 15 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{activity_id: } activity identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{activity_latitude: } activity latitude, type \code{\link[base]{numeric}}.
 \item{activity_longitude: } activity longitude, type \code{\link[base]{numeric}}.
 \item{activity_date: } activity date, type \code{\link[base]{character}}.
 \item{activity_code: } activity code to define the type of activity, type \code{\link[base]{integer}}.
 \item{ocean_code: } ocean code, type \code{\link[base]{integer}}.
  For example \code{ocean_code=1} for the Atlantic Ocean and \code{ocean_code=2} the Indian Ocean.
 \item{school_type_code:} school type code, type \code{\link[base]{integer}}.
  In Observe referential template: 1 for floating object school, 2 for free school and 0 for undetermined school.
 \item{positive_set_count: } count of positive set (catch weight and/or catch count not zero), type \code{\link[base]{integer}}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-fishing_effort"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-fishing_effort}{}}}
\subsection{Method \code{fishing_effort()}}{
Process for set duration, time at sea, fishing time and searching time calculation (in hours).
Details about the methods are available in the vignette : \href{https://ob7-ird.github.io/t3/articles/level_1.html#set-duration-calculation}{Process 1.4: Fishing effort indicators calculation}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$fishing_effort(
  set_duration_ref,
  activity_code_ref,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  global_output_path = NULL,
  referential_template = "observe"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_duration_ref}}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Data and parameters for set duration calculation (by year, country, ocean and school type), in the same format as the \href{https://ob7-ird.github.io/t3/reference/set_duration_ref.html}{referential set duration table}.
Duration in minutes in the reference table, converted into hours in output for subsequent processing).}

\item{\code{activity_code_ref}}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Reference table with the activity codes to be taken into account for the allocation of sea and/or fishing time,
 and/or searching time and/or set duration.}

\item{\code{sunrise_schema}}{Object of class {\link[base]{character}} expected. Sunrise characteristic. By default "sunrise" (top edge of the sun appears on the horizon). See below for more details.}

\item{\code{sunset_schema}}{Object of class {\link[base]{character}} expected. Sunset characteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts).
See below for more details.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}} expected if parameter outputs_extraction equal TRUE.
Path of the global outputs directory. The function will create subsection if necessary.
 By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}

\item{\code{referential_template}}{Object of class \code{\link[base]{character}} expected. By default "observe".
Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Available variables are:
\itemize{
 \item{"sunrise": } sunrise (top edge of the sun appears on the horizon)
 \item{"sunriseEnd": } sunrise ends (bottom edge of the sun touches the horizon)
 \item{"goldenHourEnd": } morning golden hour ends(soft light, best time for photography)
 \item{"solarNoon": } solar noon (sun is in the highest position)
 \item{"goldenHour": } evening golden hour starts
 \item{"sunsetStart": } sunset starts (bottom edge of the sun touches the horizon)
 \item{"sunset": } sunset (sun disappears below the horizon, evening civil twilight starts)
 \item{"dusk": } dusk (evening nautical twilight starts)
 \item{"nauticalDusk": } nautical dusk (evening astronomical twilight starts)
 \item{"night": } night starts (dark enough for astronomical observations)
 \item{"nadir": } nadir (darkest moment of the night, sun is in the lowest position)
 \item{"nightEnd": } night ends (morning astronomical twilight starts)
 \item{"nauticalDawn": } nautical dawn (morning nautical twilight starts)
 \item{"dawn": } dawn (morning nautical twilight ends, morning civil twilight starts)
 }
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level1/data/". \cr
 process_1_4: a table (.csv) with as many rows as activities and 20 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{activity_id: } activity identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{activity_latitude: } activity latitude, type \code{\link[base]{numeric}}.
 \item{activity_longitude: } activity longitude, type \code{\link[base]{numeric}}.
 \item{activity_date: } activity date, type \code{\link[base]{character}}.
 \item{activity_code: } activity code to define the type of activity, type \code{\link[base]{integer}}.
 \item{objectoperation_code: } object operation code to define the type of floating object operation (in Observe referential), type \code{\link[base]{character}}.
 \item{ocean_code: } ocean code, type \code{\link[base]{integer}}.
  For example \code{ocean_code=1} for the Atlantic Ocean and \code{ocean_code=2} the Indian Ocean.
 \item{school_type_code:} school type code, type \code{\link[base]{integer}}.
  In Observe referential template: 1 for floating object school, 2 for free school and 0 for undetermined school.
 \item{positive_set_count: } count of positive set (catch weight and/or catch count not zero), type \code{\link[base]{integer}}.
 \item{set_duration: } set duration in hours, according to the \href{https://ob7-ird.github.io/t3/reference/set_duration_ref.html}{referential set duration table}, type \code{\link[base]{numeric}}.
 \item{time_at_sea: } time at sea in hours, type \code{\link[base]{numeric}}.
 \item{fishing_time: } fishing time in hours, type \code{\link[base]{numeric}}.
 \item{searching_time: } searching time in hours, type \code{\link[base]{numeric}}.\cr
  Equal to the fishing time value minus the sum of the sets duration values.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_length_class_ld1_to_lf"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_length_class_ld1_to_lf}{}}}
\subsection{Method \code{sample_length_class_ld1_to_lf()}}{
Process for length conversion, if necessary, in length fork (lf). \cr
In fact, during the sampling process, fishes length can be collected and expressed in different standards. \cr
For example, regarding field constraints and more precisely the length of the different species, sampling data covered in T3 can by express in first dorsal length (LD1) or curved fork length (LF).\cr
 Generally, length of small individuals are provided in LF because it's logistically possible and easier to measure the entire fish, while length of bigger individuals are provided in LD1, for the same reciprocal reasons.
This step aims to standardize this standard among sampling data and at the end have only length sampling data expressed in LF. \cr
Historical and so far, the process use a referential \href{https://ob7-ird.github.io/t3/reference/length_step.html}{conversion table LD1 to LF}.
In addition, the \code{sample_number_measured} variable, in this step will be converted to a \code{sample_number_measured_lf} variable (notably due to the creation of new samples to split one LD1 class in multiples LF classes during certain conversions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_ld1_to_lf(
  length_step,
  global_output_path = NULL,
  referential_template = "observe"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length_step}}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Data frame object with length ratio between ld1 and lf class, in the same format as the \href{https://ob7-ird.github.io/t3/reference/length_step.html}{conversion table LD1 to LF}.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}

\item{\code{referential_template}}{Object of class \code{\link[base]{character}} expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_1: a table (.csv) with as many rows as elementary samples raw, plus the elementary samples raw created by certain conversions from LD1 TO LF classes, and 16 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{sub_sample_id: } sub-sample identification number, type \code{\link[base]{integer}}.
 \item{elementarysampleraw_id: } elementarysampleraw identification (unique topiaid from database (ps_logbook.samplespeciesmeasure in Observe)), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_length_class: } sample length class (cm) of measured individuals in first dorsal length (LD1), type \code{\link[base]{numeric}}.
 \item{sample_number_measured: } sample number of measured individuals in first dorsal length (LD1), type \code{\link[base]{integer}}.
 \item{sample_length_class_lf: } sample length class  (cm) of measured individuals converted in curved fork length (LF), type \code{\link[base]{numeric}}.
 \item{sample_number_measured_lf: } sample number of measured individuals converted for curved fork length (LF) distribution, type \code{\link[base]{numeric}}.
 For example, for one sample (\code{sample_number_measured=1}) from the Atlantic Ocean (1), of the species YFT (\emph{Thunnus albacares}), with a first dorsal length class (LD1) measured at \code{sample_length_class=8} (cm),
 the LD1 to LF conversion will create a new elementary sample row because a percentage of \code{ratio=50} \% of the number of fish in the sample will be assigned to the curved fork length class: \code{sample_number_measured_lf=32} (cm) ( \code{sample_number_measured_lf=0.5})
 and 50\% of this sample will obtain a \code{sample_number_measured_lf=34} (cm) with \code{sample_number_measured_lf=0.5}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_number_measured_extrapolation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_number_measured_extrapolation}{}}}
\subsection{Method \code{sample_number_measured_extrapolation()}}{
Process for sample number measured individuals extrapolation to sample number individuals counted.
In fact, during the sampling and according to the protocol, just a part of the sampled individuals are measured in relation to that counted.
The aim of this step is to extrapolate the number of individuals measured in the sample to the number of individuals counted in the sample.
To do that, a Raising Factor (RF4) is calculated per stratum, per well, per sample, per sub-sample and per species.
It is equal, by stratum, to the sum of each counted individuals divided by the sum of measured individuals (after conversion of measure in curved fork length in the process 2.1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_number_measured_extrapolation(global_output_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_2: a table (.csv) with as many rows as elementary samples raw, plus the elementary samples raw created by certain conversions from LD1 TO LF classes, and 17 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{sub_sample_id: } sub-sample identification number, type \code{\link[base]{integer}}.
 \item{sub_sample_total_count_id: } sub sample identification bis in relation with the fish total count (unique topiaid from database (ps_logbook.samplespecies in Observe)), type \code{\link[base]{character}}.
 \item{elementarysampleraw_id: } elementarysampleraw identification (unique topiaid from database (ps_logbook.samplespeciesmeasure in Observe)), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_length_class_lf: } sample length class (cm) of measured individuals converted in curved fork length (LF), type \code{\link[base]{numeric}}.
 \item{sample_number_measured_lf: } sample number of measured individuals converted for curved fork length (LF) distribution, type \code{\link[base]{numeric}}.
 \item{sample_total_count: } total number of individuals counted for this sample, type \code{\link[base]{integer}}.
 \item{sample_number_measured_extrapolated_lf: } sample number of measured individuals (converted in LF) extrapolated to the sample number of counted individuals, type \code{\link[base]{numeric}}.\cr
 \code{sample_number_measured_extrapolated_lf=sample_number_measured_lf x rf4}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-sample_length_class_step_standardisation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-sample_length_class_step_standardisation}{}}}
\subsection{Method \code{sample_length_class_step_standardisation()}}{
Process for step standardisation of lf length class.
 This step aims to standardize sample length classes. So far, these specifications are integrate in the process:
 \itemize{
\item{ a length classes step of 1cm for: } SKJ (\emph{Katsuwonus pelamis}), LTA (\emph{Euthynnus alletteratus}) and FRI (\emph{Auxis thazard}),
\item{ a length classes step of 2cm fo: } YFT (\emph{Thunnus albacares}), BET (\emph{Thunnus obesus}) and ALB (\emph{Thunnus alalunga}).
}
To standardize the original sample's curved fork length (LF), the object "elementarysample" is created by  aggregation of elementary sample raw.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$sample_length_class_step_standardisation(
  maximum_lf_class = as.integer(500),
  global_output_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maximum_lf_class}}{Object of type \code{\link[base]{integer}} expected. Theoretical maximum lf class that can occur (all species considerated). By default 500.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory.The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_3: a table (.csv) with as many rows as elementary samples, and 17 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{sample_type_code: } sample type type code, type \code{\link[base]{integer}}.
 \item{sample_quality_code: } sample quality code, type \code{\link[base]{integer}}.
 \item{sub_sample_id: } sub-sample identification number, type \code{\link[base]{integer}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_total_count: } total number of individuals counted for this sample, type \code{\link[base]{integer}}.
 \item{sample_standardised_length_class_lf: } standardised sample length class (cm) in curved fork length (LF), according to the species and step associated, type \code{\link[base]{numeric}}.
 \item{sample_number_measured_extrapolated_lf: } standardised sample number of measured individuals (converted in LF and extrapolated in step \href{https://ob7-ird.github.io/t3/articles/level_2.html#process-2-2-sample-number-measured-extrapolation}{2.2}, type \code{\link[base]{numeric}}) .\cr
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-well_set_weight_categories"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-well_set_weight_categories}{}}}
\subsection{Method \code{well_set_weight_categories()}}{
Process for well set weight categories definition.
The sampling provides information at the well scale. However, a set can be split between several wells and the individuals sampled could belong to as many sets as there are in the well.
This process aim to compute a weighted weight, which represents the weight of a set in a well, according to the distribution of this set in all the wells.\cr
The overall formula is as follows:
\eqn{WW = \frac{W1}{W2} \times WT}{WW=W1/W2 x WT}, where:
\itemize{
 \item{WW: } is the weighted weight,
 \item{W1: } is the weight of the set in the well,
 \item{W2: } is the weight of the set in all the sampled wells,
 \item{WT: } the total set's weight.
  }
So far, the process is developed for the purse seiner.
Furthermore, a proportion of each sampling sets among the sampling well will be calculated in relation with the weighted weight:\cr
\eqn{PWW = \frac{WW_{i,j}}{\sum_{i=1}^{n} WW_{i,j}}}{WW_ij / sum_i(WW_ij)}, where:
\itemize{
 \item{PWW: } is the proportional weighted weight,
 \item{\eqn{WW_{i,j}}{WW_ij}: } is the weighted weight of the current set i in well j.
 }
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$well_set_weight_categories(
  sample_set,
  global_output_path = NULL,
  referential_template = "observe"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_set}}{Object of type \code{\link[base]{data.frame}} expected. Data frame object with weighted weigh of each set sampled.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}

\item{\code{referential_template}}{Object of class \code{\link[base]{character}} expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_4: a table (.csv) with as many rows as elementary samples, and 12 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{activity_id: } activity identification (unique topiaid from database (ps_logbook.activity in Observe)), type \code{\link[base]{character}}.
 \item{weighted_weight_minus10: } weighted catch weight of individuals in the less than 10 tonnes category (by well, in tonnes, considering all species), type \code{\link[base]{numeric}}.
 \item{weighted_weight_plus10: } weighted catch weight of individuals in the over 10 kg category  (by well, in tonnes, considering all species), type \code{\link[base]{numeric}}.
 \item{weighted_weight: } weighted catch weight (WW) of individuals (less and more 10kg categories, by well, in tonnes, considering all species), which represents the weight of a set in a well, type \code{\link[base]{numeric}}.
 }
 To better understand what the process does, let's look at an example:\cr
 A set of 90 tonnes is display in 3 wells, 40 tonnes in the first one, 30 tonnes in the second and 20 tonnes in the last one.\cr
 The wells 2 and 3 were sampled but no the first one. For the second well, the weighted weight will be equal to 54 tonnes (30 / 50 x 90). \cr
 For the third one, the weighted weight will be equal to 36 tonnes (20 / 50 x 90).
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-standardised_sample_creation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-standardised_sample_creation}{}}}
\subsection{Method \code{standardised_sample_creation()}}{
Object standardised sample creation.
This process aims to sum up the samples according to the update made from the processes 2.1 to 2.3 on sample data.
In this step we left behind all the notions of subsamples and we take into account the new-sample creation in the step above (for example when we make the conversion to LD1 to LF).
This step create a new object called standardized sample expressed at the scale of the trip, the well, the sample (id, quality and type) and the species.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_creation(global_output_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_5: a table (.csv) with as many rows as standardized samples, and 12 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_standardised_length_class_lf: } standardized sample length class in curved fork length (LF) (cm), type \code{\link[base]{numeric}}.
 \item{sample_number_measured_extrapolated_lf: } standardized sample number of measured individuals (converted in LF and  extrapolated to all counted individuals), type \code{\link[base]{numeric}}.\cr
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-standardised_sample_set_creation"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-standardised_sample_set_creation}{}}}
\subsection{Method \code{standardised_sample_set_creation()}}{
R6 object standardised sample set creation.
In the previous processes and in the object standardized sample associated, samples is expressed at the well scale.\cr
In this step, the aim is to move to the expression of sample by well to sample by set. \cr
In the process \href{https://ob7-ird.github.io/t3/articles/level_2.html#process-2-4-well-set-weight-categories}{2.4}, a weighted weight (WW) and a proportion of this weighted weight (PWW) at the set scale, has been calculated.\cr
By combination of this value and elements of the object standardized sample, a new object called standardized sample set was created. Like explain before, this object is the expression of the sample at the set scale.\cr
Furthermore, this process made a conversion of the samples length measurements in weight by length weight relationships (LWR).\cr
LWR formulas take the form: \eqn{RWT=a \times LF^b}{RWT= a x (LF)^b}, where:
\itemize{
\item{RWT: } is the round weight (kg),
\item {LF: } is the curved fork length (cm),
\item{parameters a and b} comes from a references table as the \href{https://ob7-ird.github.io/t3/reference/length_weight_relationship.html}{Referential LWR table} and are dependent of of the species and potentially of the area (ocean or others) and the season.
}
More detail information could be find on the regional fisheries management organisations (RFMOs) like \href{https://www.iccat.int/}{ICCAT} or \href{https://iotc.org/}{IOTC}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$standardised_sample_set_creation(
  length_weight_relationship_data,
  global_output_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length_weight_relationship_data}}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Data frame object with parameters for length weight relationships, in the same format as the \href{https://ob7-ird.github.io/t3/reference/length_weight_relationship.html}{Referential LWR table}.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}.
Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_6: a table (.csv) with as many rows as , and 15 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_standardised_length_class_lf: } standardized sample length class in curved fork length (LF) (cm), type \code{\link[base]{numeric}}.
 \item{sample_number_weighted: } sample number of measured individuals weighted by set weight,  after conversion in LF and extrapolation to all counted individuals, type \code{\link[base]{numeric}}.\cr
 \code{sample_number_weighted = sample_number_measured_extrapolated_lf * PWW}.
 \item{sample_weight_unit: } weight (kg) of one individual,
  calculated using length weight relationships as in \href{https://ob7-ird.github.io/t3/reference/length_weight_relationship.html}{Referential LWR table}: \code{sample_weight_unit = parameter_a * sample_standardised_length_class_lf ^ parameter_b}),
   type \code{\link[base]{numeric}}.
 \item{sample_weight: } weight (kg) of all measured individuals weighted by set weight, after conversion in LF and extrapolation to all counted individuals, type \code{\link[base]{numeric}}.\cr
 \code{sample_weight = sample_weight_unit * sample_number_weighted}.
 \item{sample_category: } sample category ("-10kg" or "+10kg"), according to \code{sample_weight_unit} value, type \code{\link[base]{character}}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-raised_factors_determination"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-raised_factors_determination}{}}}
\subsection{Method \code{raised_factors_determination()}}{
Raised factors determination for weight sample set to set.
This step aims to check relevance of the object standardized sample set by calculation of 6 parameters at the scale of each well sets, if it's possible regarding data available:
 \itemize{
 \item{number and weight of sampled individuals, total and by weight categories (\eqn{\leq}{<=} 10kg or > 10kg):
 \itemize{
  \item{\code{weighted_samples_minus10}: } sum of \code{sample_weight} for sample weight category \eqn{\leq}{<=} 10 kg, from \code{standardised_sample_set} object,
   created at step \href{https://ob7-ird.github.io/t3/reference/full_trips.html#method-standardised-sample-set-creation-}{2.6}.
   \item{\code{weighted_samples_plus10}: } sum of \code{sample_weight} for sample weight category > 10 kg, from \code{standardised_sample_set} object,
   created at step \href{https://ob7-ird.github.io/t3/reference/full_trips.html#method-standardised-sample-set-creation-}{2.6}.
    \item{\code{weighted_samples_total}: } sum of \code{sample_weight} for all sample weight categories (\eqn{\leq}{<=} 10 kg) and > 10kg, from \code{standardised_sample_set} object,
   created at step \href{https://ob7-ird.github.io/t3/reference/full_trips.html#method-standardised-sample-set-creation-}{2.6}.
 }
 }
 \item{three raising factors are calculated related to the weighted weight of the set (calculated at step \href{https://ob7-ird.github.io/t3/reference/full_trips.html#method-well-set-weight-categories-}{2.4} and weight of sampled individuals, total and by weight categories (\eqn{\leq}{<=} 10kg and > 10kg):
 \itemize{
  \item{\code{rf_minus10= weighted_weight_minus10 / weighted_samples_minus10}}
   \item{\code{rf_plus10= weighted_weight_plus10 / weighted_samples_plus10}}
    \item{\code{rf_total=weighted_weight / weighted_samples_total}}
 }
 }
 }
 The verification thresholds can be modified in the function parameters using the following arguments:
 \itemize{
 \item{\code{threshold_rf_minus10}: } by default at 500,
  \item{\code{threshold_rf_plus10}: } by default at 500,
  \item{\code{threshold_frequency_rf_minus10}: } by default at 75,
  \item{\code{threshold_frequency_rf_plus10}: } by default at 75,
  \item{\code{threshold_rf_total}: } by default at 250.
  }
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_factors_determination(
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250),
  global_output_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold_rf_minus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor on individuals category minus 10. By default 500.}

\item{\code{threshold_rf_plus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor on individuals category plus 10. By default 500.}

\item{\code{threshold_frequency_rf_minus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limit frequency value for raising factor on individuals category minus 10. By default 75.}

\item{\code{threshold_frequency_rf_plus10}}{Object of type \code{\link[base]{integer}} expected. Threshold limit frequency value for raising factor on individuals category plus 10. By default 75.}

\item{\code{threshold_rf_total}}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor (all categories). By default 250.}

\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}.
Path of the global outputs directory. The function will create subsection if necessary.
 By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_7: a table (.csv) with as many rows as , and 13 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{activity_id: } activity identification (unique topiaid from database (ps_logbook.activity in Observe)), type \code{\link[base]{character}}.
 \item{weighted_samples_minus10: } weight of sampled individuals  (tonnes), for weight category (\eqn{\leq}{<=} 10kg, type \code{\link[base]{numeric}}.
 \item{weighted_samples_plus10: } weight of sampled individuals (tonnes), for weight category > 10kg, type \code{\link[base]{numeric}}.
 \item{weighted_samples_total: } weight of sampled individuals  (tonnes), for all weight categories (\eqn{\leq}{<=} 10kg and > 10kg), type \code{\link[base]{numeric}}.
 \item{rf_validation: } raising factor status, type \code{\link[base]{integer}}.
 \item{rf_validation_label: } raising factor status label, type \code{\link[base]{character}}.
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-raised_standardised_sample_set"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-raised_standardised_sample_set}{}}}
\subsection{Method \code{raised_standardised_sample_set()}}{
Application of process 2.8 raised factors on standardised sample set.
This last step aim to express number and weight of sampled individuals at the scale of the set. The process use the factors calculated in the process \href{https://ob7-ird.github.io/t3/articles/level_2.html#process-2-7-raised-factors-determination}{2.7}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$raised_standardised_sample_set(global_output_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}. Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a global_output_path is specified, the following output is extracted and saved in ".csv" format under the path: "global_output_path/level2/data/". \cr
 process_2_8: a table (.csv) with as many rows as , and 14 columns:
 \itemize{
 \item{full_trip_id: } retained full trip id, type \code{\link[base]{integer}}.
 \item{full_trip_name: } full trip id, type \code{\link[base]{integer}}.
 \item{trip_id: } trip identification (unique topiaid from database), type \code{\link[base]{character}}.
 \item{trip_end_date: } trip end date, type \code{\link[base]{character}}.
 \item{year_trip_end_date: } year of trip end, type \code{\link[base]{integer}}.
 \item{vessel_code: } vessel code, type \code{\link[base]{integer}}.
 \item{vessel_type_code: } vessel type code, type \code{\link[base]{integer}}.
 \item{well_id: } well identification (unique topiaid from database (ps_logbook.well in Observe)), type \code{\link[base]{character}}.
 \item{activity_id: } activity identification (unique topiaid from database (ps_logbook.activity in Observe)), type \code{\link[base]{character}}.
 \item{sample_id: } sample identification (unique topiaid from database (ps_logbook.sample in Observe)), type \code{\link[base]{character}}.
 \item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
 \item{sample_standardised_length_class_lf: } standardized sample length class in curved fork length (LF), (cm), type \code{\link[base]{numeric}}.
 \item{sample_number_weighted_set: } sample number weighted by set, type \code{\link[base]{numeric}}. \cr
 \code{sample_number_weighted_set = sample_number_weighted * rf}, where :
 \itemize{
 \item{\code{rf}: } is one of the raising factors calculated in the process \href{https://ob7-ird.github.io/t3/articles/level_2.html#process-2-7-raised-factors-determination}{2.7},
  according to sample weight category (\eqn{\leq}{<=} 10kg and > 10kg).
 \item{\code{sample_number_weighted}: } is the sample number of measured individuals weighted by set weight, after conversion in LF and extrapolation to all counted individuals,
  calculated in the process \href{https://ob7-ird.github.io/t3/articles/level_2.html#process-2-6-sample-number-standardisation}{2.6}.
 }
 \item{sample_weight_set: } sample weight by set (tonnes), type \code{\link[base]{numeric}}.\cr
 \code{sample_weight_set = sample_weight_unit/1000 * sample_number_weighted_set}
 }
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-path_to_level3"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-path_to_level3}{}}}
\subsection{Method \code{path_to_level3()}}{
Temporary link to the R object model with modelisation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$path_to_level3(global_output_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{global_output_path}}{By default object of type \code{\link[base]{NULL}} but object of type \code{\link[base]{character}}.
Path of the global outputs directory. The function will create subsection if necessary.
By default NULL, for no outputs extraction. Outputs will be extracted, only if a global_output_path is specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-data_preparatory"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-data_preparatory}{}}}
\subsection{Method \code{data_preparatory()}}{
Data preparatory for the t3 modelling process (level 3).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$data_preparatory(
  inputs_level3 = NULL,
  inputs_level3_path = NULL,
  output_directory,
  periode_reference_level3 = NULL,
  target_year = as.integer(lubridate::year(Sys.time() - 1)),
  period_duration = 4L,
  target_ocean = NULL,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs_level3}}{Object of type \code{\link[base]{data.frame}} expected. Inputs of levels 3 (see function path to level 3).}

\item{\code{inputs_level3_path}}{Object of type \code{\link[base]{character}} expected. Path to the folder containing yearly data output of the level 1 and 2 (output of the function the path to level 3). If provide, replace the inputs_level3 object.}

\item{\code{output_directory}}{Object of type \code{\link[base]{character}} expected. Path of the outputs directory.}

\item{\code{periode_reference_level3}}{Object of type \code{\link[base]{integer}} expected. Year(s) period of reference for modelling estimation.}

\item{\code{target_year}}{Object of type \code{\link[base]{integer}} expected. Year of interest for the model estimation and prediction.Default value is current year -1.}

\item{\code{period_duration}}{Object of type \code{\link[base]{integer}} expected. number of years use for the modelling. The default value is 5}

\item{\code{target_ocean}}{Object of type \code{\link[base]{integer}} expected. The code of ocean of interest.}

\item{\code{distance_maximum}}{Object of type \code{\link[base]{integer}} expected. Maximum distance between all sets of a sampled well. By default 5.}

\item{\code{number_sets_maximum}}{Object of type \code{\link[base]{integer}} expected. Maximum number of sets allowed in mixture. By default 5.}

\item{\code{set_weight_minimum}}{Object of type \code{\link[base]{integer}} expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.}

\item{\code{minimum_set_frequency}}{Object of type \code{\link[base]{numeric}} expected. Minimum threshold proportion of set in a well to be used for model training in the process. By default 0.1.}

\item{\code{vessel_id_ignored}}{Object of type \code{\link[base]{integer}} expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-random_forest_models"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-random_forest_models}{}}}
\subsection{Method \code{random_forest_models()}}{
Modelling proportions in sets througth random forest models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$random_forest_models(
  output_level3_process1,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_level3_process1}}{Object of type \code{\link[base]{data.frame}} expected. Output table data_lb_sample_screened from process 3.1.}

\item{\code{num.trees}}{Object of type \code{\link[base]{integer}} expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.}

\item{\code{mtry}}{Object of type \code{\link[base]{integer}} expected. Number of variables randomly sampled as candidates at each split. The default value is 2.}

\item{\code{min.node.size}}{Object of type \code{\link[base]{numeric}} expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.}

\item{\code{seed_number}}{Object of type \code{\link[base]{integer}} expected. Set the initial seed for the modelling. The default value is 7.}

\item{\code{small_fish_only}}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-models_checking"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-models_checking}{}}}
\subsection{Method \code{models_checking()}}{
Load each full model and compute figure and tables to check the model quality. Furthermore, create a map of samples used for each model and relationship between logbook reports and samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$models_checking(
  output_level3_process2,
  output_directory,
  plot_sample = FALSE,
  avdth_patch_coord = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_level3_process2}}{Object of type \code{\link[base]{list}} expected. Outputs models and data from process 3.2.}

\item{\code{output_directory}}{Object of type \code{\link[base]{character}} expected. Outputs directory path.}

\item{\code{plot_sample}}{\code{\link[base]{logical}}. Whether the sample figure is computed. Default value = F}

\item{\code{avdth_patch_coord}}{parameter waiting for coordinate conversion patch from avdth database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-data_formatting_for_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-data_formatting_for_predictions}{}}}
\subsection{Method \code{data_formatting_for_predictions()}}{
Formatting data for model predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$data_formatting_for_predictions(
  inputs_level3,
  output_level3_process1,
  target_year,
  vessel_id_ignored = NULL,
  country_flag = NULL,
  input_type = "observe_database",
  small_fish_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs_level3}}{Object of type \code{\link[base]{data.frame}} expected. Inputs of levels 3 (see function path to level 3).}

\item{\code{output_level3_process1}}{Object of type \code{\link[base]{data.frame}} expected. Output table data_lb_sample_screened from process 3.1.}

\item{\code{target_year}}{Object of type \code{\link[base]{integer}} expected. The year of interest for the model estimation and prediction.}

\item{\code{vessel_id_ignored}}{Object of type \code{\link[base]{integer}} expected. Specify here vessel(s) id(s) if you want to ignore it in the model estimation and prediction .By default NULL.}

\item{\code{country_flag}}{Three letters FAO flag code of country to estimate catches.}

\item{\code{input_type}}{Type of coding use in different databases. Default value is 'observe_database'. Values can be 'observe_database' or 'avdth_database'.}

\item{\code{small_fish_only}}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-model_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-model_predictions}{}}}
\subsection{Method \code{model_predictions()}}{
Model predictions for the species composition and computing of catches.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$model_predictions(
  output_level3_process2,
  output_level3_process4,
  output_directory,
  country_flag = NULL,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_level3_process2}}{Object of type \code{\link[base]{list}} expected. Outputs from level 3 process 2 (random forest models).}

\item{\code{output_level3_process4}}{Object of type \code{\link[base]{list}} expected. Outputs from level 3 process 4 (data formatting for predictions).}

\item{\code{output_directory}}{Object of type \code{\link[base]{character}} expected. Outputs directory path.}

\item{\code{country_flag}}{Three letters FAO flag code of country to estimate catches.}

\item{\code{ci}}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.}

\item{\code{ci_type}}{Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if  the ci parameter is TRUE.}

\item{\code{Nboot}}{Object of type \code{\link[base]{numeric}} expected. The number of bootstrap samples desired for the ci computation. The default value is 10.}

\item{\code{plot_predict}}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether maps of catch at size have to be done.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-show_me_what_you_got"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-show_me_what_you_got}{}}}
\subsection{Method \code{show_me_what_you_got()}}{
Most powerfull and "schwifty" function in the univers for "open the T3 process" and manipulate in live R6 objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$show_me_what_you_got()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_trips-clone"></a>}}
\if{latex}{\out{\hypertarget{method-full_trips-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{full_trips$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
