% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R, R/list_t3.R
\docType{data}
\name{list_t3}
\alias{list_t3}
\title{R6 class list_t3 creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
list_t3

list_t3
}
\description{
Create R6 reference object class list_t3

Create R6 reference object class list_t3
}
\keyword{datasets}
