% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_t3.R
\name{list_t3}
\alias{list_t3}
\title{R6 class list_t3 creation}
\description{
Create R6 reference object class list_t3
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{list_t3$new()}}
\item \href{#method-name}{\code{list_t3$name()}}
\item \href{#method-add}{\code{list_t3$add()}}
\item \href{#method-remove}{\code{list_t3$remove()}}
\item \href{#method-view}{\code{list_t3$view()}}
\item \href{#method-count}{\code{list_t3$count()}}
\item \href{#method-extract}{\code{list_t3$extract()}}
\item \href{#method-filter}{\code{list_t3$filter()}}
\item \href{#method-clone}{\code{list_t3$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$name(attribut = "all")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$add(new_item, attribut = "data")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$remove(item_id, attribut = "data")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$view(..., attribut = "data")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$count(attribut = "data")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$extract(attribut_l1 = "data", attribut_l2 = NULL, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$filter(attribut_l1 = "data", filter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{list_t3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
