% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_elementarycatch.R
\name{object_elementarycatch}
\alias{object_elementarycatch}
\title{R6 object object_elementarycatch creation}
\usage{
object_elementarycatch(trip_id, landing_date, activity_date,
  activity_number, logbook_category, vessel_id, specie_code3l,
  catch_weight)
}
\value{
A R6 reference object with data related to one elementary catch
}
\description{
Creation object object_elementarycatch in relation with R6 reference object class elementarycatch
}
\seealso{
\code{\link{elementarycatch}}
}
