% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model_data.R
\name{object_model_data}
\alias{object_model_data}
\title{R6 class object_model_data}
\description{
Create R6 reference object class object_model_data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-object_model_data-trips_object_creation}{\code{object_model_data$trips_object_creation()}}
\item \href{#method-object_model_data-activities_object_creation}{\code{object_model_data$activities_object_creation()}}
\item \href{#method-object_model_data-elementarycatches_object_creation}{\code{object_model_data$elementarycatches_object_creation()}}
\item \href{#method-object_model_data-elementarylandings_object_creation}{\code{object_model_data$elementarylandings_object_creation()}}
\item \href{#method-object_model_data-wells_object_creation}{\code{object_model_data$wells_object_creation()}}
\item \href{#method-object_model_data-setdurationrefs_data}{\code{object_model_data$setdurationrefs_data()}}
\item \href{#method-object_model_data-lengthsteps_data}{\code{object_model_data$lengthsteps_data()}}
\item \href{#method-object_model_data-samplesets_data}{\code{object_model_data$samplesets_data()}}
\item \href{#method-object_model_data-lengthweightrelationships_data}{\code{object_model_data$lengthweightrelationships_data()}}
\item \href{#method-object_model_data-clone}{\code{object_model_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-trips_object_creation"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-trips_object_creation}{}}}
\subsection{Method \code{trips_object_creation()}}{
Creation of an R6 reference object class trips which contains one or more R6 reference object class trip.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$trips_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  time_period = NULL,
  fleet_code = NULL,
  ocean_code = NULL,
  vessel_type_code = NULL,
  trip_id = NULL,
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".}

\item{\code{time_period}}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{\code{fleet_code}}{Object of class {\link[base]{character}} expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{vessel_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data csv/RData file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-activities_object_creation"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-activities_object_creation}{}}}
\subsection{Method \code{activities_object_creation()}}{
Creation of a R6 reference object class activities which contain one or more R6 reference object class activity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$activities_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  time_period = NULL,
  fleet_code = NULL,
  ocean_code = NULL,
  vessel_type_code = NULL,
  trip_id = NULL,
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".}

\item{\code{time_period}}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{\code{fleet_code}}{Object of class {\link[base]{character}} expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{vessel_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data csv/RData file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-elementarycatches_object_creation"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-elementarycatches_object_creation}{}}}
\subsection{Method \code{elementarycatches_object_creation()}}{
Creation of a R6 reference object class elementarycatches which contain one or more R6 reference object class elementarycatch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$elementarycatches_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  time_period = NULL,
  fleet_code = NULL,
  ocean_code = NULL,
  vessel_type_code = NULL,
  species_fate_code = NULL,
  trip_id = NULL,
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".}

\item{\code{time_period}}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{\code{fleet_code}}{Object of class {\link[base]{character}} expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{vessel_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{species_fate_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Specie fate(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data csv/RData file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-elementarylandings_object_creation"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-elementarylandings_object_creation}{}}}
\subsection{Method \code{elementarylandings_object_creation()}}{
Creation of a R6 reference object class elementarylandings which contain one or more R6 reference object class elementarylanding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$elementarylandings_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  time_period = NULL,
  fleet_code = NULL,
  ocean_code = NULL,
  vessel_type_code = NULL,
  trip_id = NULL,
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".}

\item{\code{time_period}}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{\code{fleet_code}}{Object of class {\link[base]{character}} expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{vessel_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data csv/RData file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default the first environment where data are found will be used. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-wells_object_creation"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-wells_object_creation}{}}}
\subsection{Method \code{wells_object_creation()}}{
Creation of a R6 reference object class wells which contain one or more R6 reference object class well, wellset, samples and elementarywellplan.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$wells_object_creation(
  data_source = "observe_database",
  database_connection = NULL,
  time_period = NULL,
  fleet_code = NULL,
  ocean_code = NULL,
  vessel_type_code = NULL,
  sample_type_code = NULL,
  trip_id = NULL,
  data_path_sample = NULL,
  data_path_wellplan = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch between "observe_database", "avdth_database", "csv_file" (with separator ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database" and "avdth_database".}

\item{\code{time_period}}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{\code{fleet_code}}{Object of class {\link[base]{character}} expected. By default NULL. Country(ies) code related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{ocean_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{vessel_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{\code{sample_type_code}}{Object of class {\link[base]{integer}} expected. By default NULL. Sample type identification.}

\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "time_periode", "country" or "ocean".}

\item{\code{data_path_sample}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data sql/csv file for samples.}

\item{\code{data_path_wellplan}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data sql/csv file for well plans.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default NULL. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-setdurationrefs_data"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-setdurationrefs_data}{}}}
\subsection{Method \code{setdurationrefs_data()}}{
Creation of a data frame object with parameters of set duration algorithms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$setdurationrefs_data(
  database_connection = NULL,
  data_source = "t3_db",
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. Path of the data sql/csv/RData file. By default NULL.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.}

\item{\code{db_con}}{Database connection R object expected. Mandatory argument for data source "t3_db", "avdth_db" and "sql_query".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-lengthsteps_data"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-lengthsteps_data}{}}}
\subsection{Method \code{lengthsteps_data()}}{
Creation of a data frame object with length ratio between ld1 and lf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$lengthsteps_data(
  data_source = "csv_file",
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "csv_file". Identification of data source. You can switch to "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Mandatory argument for data source "csv_file", "rdata_file" or "envir". Path of the data file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default NULL. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-samplesets_data"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-samplesets_data}{}}}
\subsection{Method \code{samplesets_data()}}{
Creation of a data frame object with weighted weigth of each set sampled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$samplesets_data(
  data_source = "t3_db",
  db_con = NULL,
  periode_reference = NULL,
  countries = NULL,
  oceans = NULL,
  data_path = NULL,
  trips_selected = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. Identification of data source. By default "t3_db" but you can switch to "sql_query", "csv" (with separator character ";" and decimal ","), "rdata" or "envir" (for an object in the R environment).}

\item{\code{db_con}}{Database connection R object expected. Mandatory argument for data source "t3_db", "avdth_db" and "sql_query".}

\item{\code{periode_reference}}{Object of class {\link[base]{integer}} expected. Year(s) of the reference period coded on 4 digits. Necessary argument for data source "t3_db". By default NULL.}

\item{\code{countries}}{Object of class {\link[base]{character}} expected. ISO code on 3 letters related to one or more countries. Necessary argument for data source "t3_db". By default NULL.}

\item{\code{oceans}}{Object of class {\link[base]{integer}} expected. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "t3_db". By default NULL.}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. Path of the data sql/csv/RData file. By default NULL.}

\item{\code{trips_selected}}{Object of class {\link[base]{character}} expected. Additional parameter only used with data source "t3_db". Use trip(s) identification(s) for selected trip(s) kept in the query (by periode of reference and countries). By default NULL.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. Specify an environment to look in for data source "envir". By default the first environment where data are found will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-lengthweightrelationships_data"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-lengthweightrelationships_data}{}}}
\subsection{Method \code{lengthweightrelationships_data()}}{
Creation of a data frame object with parameters for length weight relationship.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$lengthweightrelationships_data(
  data_source = "observe_database",
  database_connection = NULL,
  data_path = NULL,
  envir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can switch to "csv_file" (with separator character ";" and decimal ","), "rdata_file" or "envir" (for an object in the R environment).}

\item{\code{database_connection}}{Database connection R object expected. By default NULL. Mandatory argument for data source "observe_database".}

\item{\code{data_path}}{Object of class {\link[base]{character}} expected. By default NULL. Path of the data file.}

\item{\code{envir}}{Object of class {\link[base]{character}} expected. By default NULL. Specify an environment to look in for data source "envir".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-object_model_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-object_model_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{object_model_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
