% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_trip.R
\name{object_trip}
\alias{object_trip}
\title{R6 object trip creation}
\usage{
object_trip(trip_id, departure_date, landing_date, logbook_availability,
  fish_hold_empty, vessel_id)
}
\arguments{
\item{trip_id}{trip_id}

\item{departure_date}{departure_date}

\item{landing_date}{landing_date}

\item{logbook_availability}{logbook_availability}

\item{fish_hold_empty}{fish_hold_empty}

\item{vessel_id}{vessel_id}
}
\value{
A R6 reference object with data related to one trip
}
\description{
Creation object trip in relation with R6 reference object class trip
}
\seealso{
\code{\link{trip}}
}
