% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observe_logbook_program}
\alias{observe_logbook_program}
\title{Reference table defining the logbook program for Observe database to be considered in T3 process.}
\format{
observe_logbook_program: data frame with 31 rows and 4 columns:
\itemize{
\item{organism: } organism using the logbook program, type \code{\link[base]{character}}.
\item{topiaid: } topiaid of the logbook program, type \code{\link[base]{character}}.
\item{code: }  code of the logbook program, type \code{\link[base]{character}}.
\item{label1: } label of the logbook program's in English, type \code{\link[base]{character}}.
}
}
\usage{
data(observe_logbook_program, package="t3")
}
\description{
Reference table defining the logbook program for Observe database to be considered in T3 process.
}
\examples{
data(observe_logbook_program, package="t3")
 observe_logbook_program
}
\keyword{wobserve_logbook_program}
