% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisedsample.R
\name{standardisedsample}
\alias{standardisedsample}
\title{R6 class standardisedsample creation}
\description{
Create R6 reference object class standardisedsample
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{standardisedsample$new()}}
\item \href{#method-clone}{\code{standardisedsample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$new(
  trip_id,
  well_id,
  sample_id,
  sample_quality,
  sample_type,
  specie_code3l,
  sample_standardised_length_class_lf,
  sample_number_measured_extrapolated_lf,
  sample_total_count,
  elementarysample
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
