% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisedsample.R
\name{standardisedsample}
\alias{standardisedsample}
\title{R6 class standardisedsample creation}
\description{
Create R6 reference object class standardisedsample
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{standardisedsample$new()}}
\item \href{#method-clone}{\code{standardisedsample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function for R6 standardisedsample class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$new(
  trip_id,
  well_id,
  sample_id,
  sample_quality,
  sample_type,
  specie_code,
  specie_code3l,
  sample_standardised_length_class_lf,
  sample_number_measured_extrapolated_lf,
  sample_total_count,
  elementarysample
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{well_id}}{(character) Well identification.}

\item{\code{sample_id}}{(character) Sample identification.}

\item{\code{sample_quality}}{(integer) Sample quality identification.}

\item{\code{sample_type}}{(integer) Sample type identification.}

\item{\code{specie_code}}{(integer) Specie code identification.}

\item{\code{specie_code3l}}{(character) Specie code identification on 3 characters.}

\item{\code{sample_standardised_length_class_lf}}{(integer) Sample standardised length class length fork of measured individus.}

\item{\code{sample_number_measured_extrapolated_lf}}{(numeric) Sample number of measured individus extrapolated to all counted individus.}

\item{\code{sample_total_count}}{(integer) Sample number of total individus counted.}

\item{\code{elementarysample}}{(R6-elementarysample) A R6 reference object class elementarysample.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
