% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisedsampleset.R
\name{standardisedsampleset}
\alias{standardisedsampleset}
\title{R6 class standardisedsampleset creation}
\description{
Create R6 reference object class standardisedsampleset
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{standardisedsampleset$new()}}
\item \href{#method-clone}{\code{standardisedsampleset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize function for R6 standardisedsampleset class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsampleset$new(
  trip_id,
  activity_id,
  well_id,
  sample_id,
  sample_quality,
  sample_type,
  specie_code3l,
  sample_standardised_length_class_lf,
  sample_number_weighted,
  sample_weigth,
  sample_weight_unit,
  sample_category,
  standardisedsample
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{activity_id}}{(character) Activity identification.}

\item{\code{well_id}}{(character) Well identification.}

\item{\code{sample_id}}{(character) Sample identification.}

\item{\code{sample_quality}}{(integer) Sample quality identification.}

\item{\code{sample_type}}{(integer) Sample type identification.}

\item{\code{specie_code3l}}{(character) Specie code identification on 3 characters.}

\item{\code{sample_standardised_length_class_lf}}{(integer) Sample standardised length class length fork of measured individus.}

\item{\code{sample_number_weighted}}{(numeric) Sample number of measured individus extrapolated to all counted individus and weighted by set weight.}

\item{\code{sample_weigth}}{(numeric) Weight of the sample number of measured individus extrapolated to all counted individus (conversion by length weight relationship: coeficient a * (length class lf ^ coeficient b)}

\item{\code{sample_weight_unit}}{(numeric or na) Weight of one individu calculated by the length weight relationship: coeficient a * (length class lf ^ coeficient b).}

\item{\code{sample_category}}{(character) Sample category: -10kg or +10kg.}

\item{\code{standardisedsample}}{(R6-standardisedsample) A R6 reference object class standardisedsample.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsampleset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
