% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisedsampleset.R
\name{standardisedsampleset}
\alias{standardisedsampleset}
\title{R6 class standardisedsampleset creation}
\description{
Create R6 reference object class standardisedsampleset
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{standardisedsampleset$new()}}
\item \href{#method-clone}{\code{standardisedsampleset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsampleset$new(
  trip_id,
  activity_id,
  well_id,
  sample_id,
  specie_code3l,
  sample_standardised_length_class_lf,
  sample_number_weighted,
  sample_weigth,
  sample_weight_unit,
  standardisedsample
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{standardisedsampleset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
