% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3-package.R
\docType{package}
\name{t3-package}
\alias{t3}
\alias{t3-package}
\title{Tropical Tunas Treatment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The multi-species nature of tropical tuna surface fisheries highlights a series of difficulties when catches are estimated by species and size statistics. The Tropical Tuna Treatment (T3) process was initiated more than 30 years ago and is still today the result of a strong collaboration between several countries (notably France, Spain, the Seychelles and Senegal). Is aims are to correct biases of the logbook data on species composition and provide more accurate catch estimations per species for the purse seiner fleet, especially European and Seychelles fleet.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/t3}
  \item \url{https://ob7-ird.github.io/t3}
  \item Report bugs at \url{https://github.com/OB7-IRD/t3/issues}
}

}
\author{
\strong{Maintainer}: Jeanne Clément \email{jeanne.clement@ird.fr} (\href{https://orcid.org/0000-0002-5228-5015}{ORCID})

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})
  \item Antoine Duparc \email{antoine.duparc@ird.fr} (\href{https://orcid.org/0000-0001-6135-9015}{ORCID})
  \item Pascal Cauquil \email{pascal.cauquil@ird.fr} (\href{https://orcid.org/0000-0003-0718-8950}{ORCID})
  \item Laurent Floc'h \email{laurent.floch@ird.fr} (\href{https://orcid.org/0000-0003-4764-9766}{ORCID})
  \item Julien Lebranchu \email{julien.lebranchu@ird.fr} (\href{https://orcid.org/0000-0002-7759-0534}{ORCID})
}

Other contributors:
\itemize{
  \item Pascal Bach \email{pascal.bach@ird.fr} (\href{https://orcid.org/0000-0003-0804-8466}{ORCID}) [contributor]
  \item Alain Fonteneau \email{alain.fonteneau@ird.fr} [contributor]
  \item Viveca Fonteneau \email{vivinord@wanadoo.fr} [contributor]
  \item Norbet Billet \email{norbert.billet@ifremer.fr} [contributor]
  \item Laurent Dubroca \email{laurent.dubroca@ifremer.fr} [contributor]
  \item Emmanuel Chassot \email{emmanuel.Chassot@fao.org} (\href{https://orcid.org/0000-0002-6779-4955}{ORCID}) [contributor]
  \item IRD - Ob7 Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
