% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3-package.R
\docType{package}
\name{t3-package}
\alias{t3}
\alias{t3-package}
\title{Tropical Tunas Treatment (T3)}
\description{
The multi-species nature of tropical tuna surface fisheries gives rise to a series of difficulties when estimating the catch by species and catch at size statistics. The T3 processing was built about 30 years ago in order to correct biases of the logbook data on species composition and to provide more accurate catch estimates per species for the European purse seine fleet.
}
\section{Process for data model initialisation}{

\itemize{
\item{\code{\link{data_model_initialisation}}}
}
}

\section{Process for level 1 (logbooks standardisation)}{

\itemize{
\item{\code{\link{t3_level1}}}
}
}

\section{Process for level 2 (samples standardisation)}{

\itemize{
\item{\code{\link{t3_level2}}}
}
}

\section{Generic process fuction}{

\itemize{
\item{\code{\link{t3_process}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/t3}
  \item \url{https://ob7-ird.github.io/t3}
  \item Report bugs at \url{https://github.com/OB7-IRD/t3/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Authors:
\itemize{
  \item Antoine Duparc \email{antoine.duparc@ird.fr} (\href{https://orcid.org/0000-0001-6135-9015}{ORCID})
  \item Pascal Cauquil \email{pascal.cauquil@ird.fr} (\href{https://orcid.org/0000-0003-0718-8950}{ORCID})
  \item Laurent Floc'h \email{laurent.floch@ird.fr} (\href{https://orcid.org/0000-0003-4764-9766}{ORCID})
  \item Julien Lebranchu \email{julien.lebranchu@ird.fr} (\href{https://orcid.org/0000-0002-7759-0534}{ORCID})
}

Other contributors:
\itemize{
  \item Pascal Bach \email{pascal.bach@ird.fr} (\href{https://orcid.org/0000-0003-0804-8466}{ORCID}) [contributor]
  \item Alain Fonteneau \email{alain.fonteneau@ird.fr} [contributor]
  \item Viveca Fonteneau \email{vivinord@wanadoo.fr} [contributor]
  \item Norbet Billet \email{norbert.billet@ifremer.fr} [contributor]
  \item Laurent Dubroca \email{laurent.dubroca@ifremer.fr} [contributor]
  \item Emmanuel Chassot \email{emmanuel.Chassot@fao.org} (\href{https://orcid.org/0000-0002-6779-4955}{ORCID}) [contributor]
  \item IRD - Ob7 Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [contributor, funder]
}

}
\keyword{internal}
