% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3-package.R
\docType{package}
\name{t3-package}
\alias{t3}
\alias{t3-package}
\title{t3 package}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\section{Class declaration}{

\itemize{
\item{\code{\link{}}}
}
}

\section{Objets creation}{

\itemize{
\item{\code{\link{}}}
}
}

\section{Functions}{

\itemize{
\item{\code{\link{}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/t3}
  \item Report bugs at \url{https://github.com/OB7-IRD/t3/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr}

Authors:
\itemize{
  \item name_Bob1 surname_bob1 \email{bob1@ird.fr}
}

Other contributors:
\itemize{
  \item name_Bob2 surname_bob2 \email{bob2@ird.fr} [copyright holder]
}

}
\keyword{internal}
