% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_level1.R
\name{t3_level1}
\alias{t3_level1}
\title{T3 process level 1}
\usage{
t3_level1(
  object_model_data,
  object_full_trips,
  rf1_computation = TRUE,
  apply_rf1_on_bycatch = FALSE,
  species_fao_codes_rf1 = c("YFT", "SKJ", "BET", "ALB", "TUN", "MIX", "LOT"),
  species_fate_codes_rf1 = as.integer(6),
  vessel_type_codes_rf1 = as.integer(c(4, 5, 6)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  fishing_effort_computation = TRUE,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  log_file = FALSE,
  log_path = NULL,
  log_name = "t3_level1",
  output_path = NULL,
  output_format = "eu",
  new_directory = FALSE,
  integrated_process = FALSE,
  referential_template = "observe"
)
}
\arguments{
\item{object_model_data}{Object of classes \code{\link[t3]{object_model_data}} and \code{\link[R6]{R6}} expected.}

\item{object_full_trips}{Object of classes \code{\link[t3]{full_trips}}, \code{\link[t3]{list_t3}} and \code{\link[R6]{R6}} expected.}

\item{rf1_computation}{Object of class \code{\link[base]{logical}} expected. If FALSE rf1 is not calculated (rf1=1 for all trips).
By default TRUE, the rf1 is calculated for each trip.}

\item{apply_rf1_on_bycatch}{Object of class \code{\link[base]{logical}} expected. By default FALSE, only the catch weights of species belonging to the species list, defined by the \code{species_fao_codes_rf1} argument are corrected, rf1 is not applied to by-catch species.
If TRUE, rf1 values will be applied to all the logbook catches associated to the trip, including by-catch species.}

\item{species_fao_codes_rf1}{Object of type \code{\link[base]{character}} expected.Specie(s) FAO code(s) used for the RF1 process.
By default, use codes YFT (*Thunnus albacares*), SKJ (*Katsuwonus pelamis*), BET (*Thunnus obesus*), ALB (*Thunnus alalunga*),
LOT (*Thunnus tonggol*) and TUN/MIX (mix of tunas species in Observe/AVDTH database) (French and Mayotte fleets).}

\item{species_fate_codes_rf1}{Object of type \code{\link[base]{integer}} expected. By default 6 ("Retained, presumably destined for the cannery"). Specie(s) fate code(s) used for the RF1 process.}

\item{vessel_type_codes_rf1}{Object of type \code{\link[base]{integer}} expected. By default 4, 5 and 6. Vessel type(s).}

\item{rf1_lowest_limit}{Object of type \code{\link[base]{numeric}} expected. Verification value for the lowest limit of the RF1. By default 0.8.}

\item{rf1_highest_limit}{Object of type \code{\link[base]{numeric}} expected. Verification value for the highest limit of the RF1. By default 1.2.}

\item{fishing_effort_computation}{Object of class {\link[base]{logical}} expected. Calculation or not of fishing effort indicators (time at sea, fishing time, set duration and searching time). By default TRUE (yes).}

\item{sunrise_schema}{Object of class {\link[base]{character}} expected. Sunrise characteristic. By default "sunrise" (top edge of the sun appears on the horizon). See function fishing_time() for more details.}

\item{sunset_schema}{Object of class {\link[base]{character}} expected. Sunset characteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See function fishing_time() for more details.}

\item{log_file}{Object of class {\link[base]{logical}} expected. Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{Object of class {\link[base]{character}} expected. Path of the log file directory. By default NULL.}

\item{log_name}{Object of class {\link[base]{character}} expected. Name of the log file. By default "t3_level1".}

\item{output_path}{Object of class \code{\link[base]{character}} expected. Outputs path directory. By default NULL.}

\item{output_format}{Object of class \code{\link[base]{character}} expected. By default "eu". Select outputs format regarding European format (eu) or United States format (us).}

\item{new_directory}{Object of class \code{\link[base]{logical}} expected. Initiate a new outputs directory of use an existing one. By default NULL.}

\item{integrated_process}{Object of class \code{\link[base]{logical}} expected. Indicate if the process is integrated in another (like the one in the function "t3_process"). By default FALSE.}

\item{referential_template}{Object of class \code{\link[base]{character}} expected. By default "observe". Referential template selected (for example regarding the activity_code). You can switch to "avdth".}
}
\value{
The function a R6 reference object of class "object_full_trips".
}
\description{
Level 1 of t3 process (logbooks standardisation).
}
