% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_process.R
\name{t3_process}
\alias{t3_process}
\title{T3 process compilation}
\usage{
t3_process(
  process = "all",
  data_source = "observe_database",
  database_connection,
  log_file = FALSE,
  log_path = NULL,
  output_path = NULL,
  new_directory = TRUE,
  years_period,
  flag_codes,
  ocean_codes,
  vessel_type_codes,
  species_fate_codes,
  sample_type_codes,
  trip_ids = NULL,
  weight_categories_avdth_ref = NULL,
  rf1_computation = TRUE,
  apply_rf1_on_bycatch = TRUE,
  species_fao_codes_rf1 = c("YFT", "SKJ", "BET", "ALB", "MIX", "TUN", "LOT"),
  species_fate_codes_rf1 = as.integer(6),
  vessel_type_codes_rf1 = as.integer(c(4, 5, 6)),
  rf1_lowest_limit = 0.8,
  rf1_highest_limit = 1.2,
  fishing_effort_computation = TRUE,
  sunrise_schema = "sunrise",
  sunset_schema = "sunset",
  maximum_lf_class = as.integer(500),
  threshold_rf_minus10 = as.integer(500),
  threshold_rf_plus10 = as.integer(500),
  threshold_frequency_rf_minus10 = as.integer(75),
  threshold_frequency_rf_plus10 = as.integer(75),
  threshold_rf_total = as.integer(250),
  periode_reference_level3 = NULL,
  target_year = as.integer(lubridate::year(Sys.time() - 1)),
  target_ocean = NULL,
  country_flag = flag_codes[1],
  period_duration = 4L,
  distance_maximum = as.integer(5),
  number_sets_maximum = as.integer(5),
  set_weight_minimum = as.integer(6),
  minimum_set_frequency = 0.1,
  vessel_id_ignored = NULL,
  num.trees = 1000L,
  mtry = 2L,
  min.node.size = 5,
  seed_number = 7L,
  small_fish_only = FALSE,
  plot_sample = FALSE,
  avdth_patch_coord = FALSE,
  ci = FALSE,
  ci_type = "all",
  Nboot = 50,
  plot_predict = FALSE
)
}
\arguments{
\item{process}{Object of class {\link[base]{character}} expected. Specify here if you want to run the whole process or just a part of it. By default "all". Check detail section below for more information.}

\item{data_source}{Object of class {\link[base]{character}} expected. By default "observe_database". Identification of data source. You can to "avdth_database".}

\item{database_connection}{Database connection, list of one or more R object(s) expected. By default NULL.
Mandatory argument for data source "observe_database" ("PostgreSQLConnection" R object), corresponding to the second element of the object returned by \href{https://ob7-ird.github.io/furdeb/reference/postgresql_dbconnection.html}{`furdeb::postgresql_dbconnection()`}.
Or mandatory argument for data source"avdth_database" ("JDBCConnection" R object) corresponding to the second element of the object returned by \href{https://ob7-ird.github.io/furdeb/reference/access_dbconnection.html}{`furdeb::access_dbconnection()`}.
For data source "observe_database", a list of "PostgreSQLConnection" R objects can be specified to query data from different observe databases.
For example, a list of two database connection arguments for "observe_main" and "observe_acquisition" can be specified to simultaneously import and process recent data from acquisition database, which has not yet been imported into the main database, and older data from the main database.}

\item{log_file}{Object of class {\link[base]{logical}} expected. Initiation or not for log file creation. By default FALSE (no).}

\item{log_path}{Object of class {\link[base]{character}} expected. Path of the log file directory. By default NULL.}

\item{output_path}{Object of class \code{\link[base]{character}} expected. Outputs path directory. By default NULL.}

\item{new_directory}{Object of class \code{\link[base]{logical}} expected. Initiate a new outputs directory of use an existing one. By default NULL.}

\item{years_period}{Object of class {\link[base]{integer}} expected. By default NULL. Year(s) of the reference time period coded on 4 digits. Mandatory for data source "observe_database" and "avdth_database".}

\item{flag_codes}{Object of class {\link[base]{character}} expected. By default NULL. Three letters country(ies) FAO code(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{ocean_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Ocean(s) related to data coded on 1 digit. Necessary argument for data source "observe_database" and "avdth_database".}

\item{vessel_type_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Vessel type(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{species_fate_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Specie fate(s) related to data extraction. Necessary argument for data source "observe_database" and "avdth_database".}

\item{sample_type_codes}{Object of class {\link[base]{integer}} expected. By default NULL. Sample type identification.}

\item{trip_ids}{Object of class {\link[base]{character}} expected. By default NULL. Additional parameter only used with data source "observe_database". Use trip(s) identification(s) for selected trip(s) kept in the query. This argument overrides all others arguments like "years_period", "country" or "ocean".}

\item{weight_categories_avdth_ref}{Object of type \code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}} expected.
Reference table defining the logbook weight categories with the following columns:
\itemize{
\item{avdth_weight_category_code: } Logbook weight category code from avdth database, type \code{\link[base]{integer}}.
\item{species_fao_code: } species FAO code, type \code{\link[base]{character}}.
\item{weight_category_min: } lower limit of the weight category (kg), type \code{\link[base]{numeric}}.
\item{weight_category_max: } upper limit of the weight category (kg), type \code{\link[base]{numeric}}.
}
Mandatory for \code{referential_template = "avdth"}.}

\item{rf1_computation}{Object of class \code{\link[base]{logical}} expected. If FALSE rf1 is not calculated (rf1=1 for all trips).
By default TRUE, the rf1 is calculated for each trip.}

\item{apply_rf1_on_bycatch}{Object of class \code{\link[base]{logical}} expected. By default FALSE, only the catch weights of species belonging to the species list, defined by the \code{species_fao_codes_rf1} argument are corrected, rf1 is not applied to by-catch species.
If TRUE, rf1 values will be applied to all the logbook catches associated to the trip, including by-catch species.}

\item{species_fao_codes_rf1}{Object of type \code{\link[base]{character}} expected.Specie(s) FAO code(s) used for the RF1 process.
By default, use codes YFT (*Thunnus albacares*), SKJ (*Katsuwonus pelamis*), BET (*Thunnus obesus*), ALB (*Thunnus alalunga*),
LOT (*Thunnus tonggol*) and TUN/MIX (mix of tunas species in Observe/AVDTH database) (French and Mayotte fleets).}

\item{species_fate_codes_rf1}{Object of type \code{\link[base]{integer}} expected. By default 6 ("Retained, presumably destined for the cannery"). Specie(s) fate code(s) used for the RF1 process.}

\item{vessel_type_codes_rf1}{Object of type \code{\link[base]{integer}} expected. By default 4, 5 and 6. Vessel type(s).}

\item{rf1_lowest_limit}{Object of type \code{\link[base]{numeric}} expected. Verification value for the lowest limit of the RF1. By default 0.8.}

\item{rf1_highest_limit}{Object of type \code{\link[base]{numeric}} expected. Verification value for the highest limit of the RF1. By default 1.2.}

\item{fishing_effort_computation}{Object of class {\link[base]{logical}} expected. Calculation or not of fishing effort indicators (time at sea, fishing time, set duration and searching time). By default TRUE (yes).}

\item{sunrise_schema}{Object of class {\link[base]{character}} expected. Sunrise characteristic. By default "sunrise" (top edge of the sun appears on the horizon). See function fishing_time() for more details.}

\item{sunset_schema}{Object of class {\link[base]{character}} expected. Sunset characteristic. By default "sunset" (sun disappears below the horizon, evening civil twilight starts). See function fishing_time() for more details.}

\item{maximum_lf_class}{Object of type \code{\link[base]{integer}} expected. Theorical maximum lf class that can occur (all species considerated). By default 500.}

\item{threshold_rf_minus10}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor on individuals category minus 10. By default 500.}

\item{threshold_rf_plus10}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor on individuals category plus 10. By default 500.}

\item{threshold_frequency_rf_minus10}{Object of type \code{\link[base]{integer}} expected. Threshold limit frequency value for raising factor on individuals category minus 10. By default 75.}

\item{threshold_frequency_rf_plus10}{Object of type \code{\link[base]{integer}} expected. Threshold limit frequency value for raising factor on individuals category plus 10. By default 75.}

\item{threshold_rf_total}{Object of type \code{\link[base]{integer}} expected. Threshold limit value for raising factor (all categories). By default 250.}

\item{periode_reference_level3}{Object of type \code{\link[base]{integer}} expected. Year(s) period of reference for modelling estimation.}

\item{target_year}{Object of type \code{\link[base]{integer}} expected. Year of interest for the model estimation and prediction.Default value is current year -1.}

\item{target_ocean}{Object of type \code{\link[base]{integer}} expected. The code of ocean of interest.}

\item{country_flag}{Three letters FAO flag code of country to estimate catches. By default, first of `flag_codes`.}

\item{period_duration}{Object of type \code{\link[base]{integer}} expected. number of years use for the modelling. The default value is 5}

\item{distance_maximum}{Object of type \code{\link[base]{integer}} expected. Maximum distance between all sets of a sampled well. By default 5.}

\item{number_sets_maximum}{Object of type \code{\link[base]{integer}} expected. Maximum number of sets allowed in mixture. By default 5.}

\item{set_weight_minimum}{Object of type \code{\link[base]{integer}} expected. Minimum set size considered. Remove smallest set for which sample could not be representative. By default 6 t.}

\item{minimum_set_frequency}{Object of type \code{\link[base]{numeric}} expected. Minimum threshold proportion of set in a well to be used for model training in the process. By default 0.1.}

\item{vessel_id_ignored}{Object of type \code{\link[base]{integer}} expected. Specify list of vessel(s) id(s) to be ignored in the model estimation and prediction .By default NULL.}

\item{num.trees}{Object of type \code{\link[base]{integer}} expected. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. The default value is 1000.}

\item{mtry}{Object of type \code{\link[base]{integer}} expected. Number of variables randomly sampled as candidates at each split. The default value is 2.}

\item{min.node.size}{Object of type \code{\link[base]{numeric}} expected. Minimum size of terminal nodes. Setting this number larger causes smaller trees to be grown (and thus take less time).The default value is 5.}

\item{seed_number}{Object of type \code{\link[base]{integer}} expected. Set the initial seed for the modelling. The default value is 7.}

\item{small_fish_only}{Object of type \code{\link[base]{logical}} expected. Whether the model estimate proportion for small fish only (< 10 kg).}

\item{plot_sample}{\code{\link[base]{logical}}. Whether the sample figure is computed. Default value = F}

\item{avdth_patch_coord}{parameter waiting for coordinate conversion patch from avdth database}

\item{ci}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether confidence interval is computed. The default value is FALSE as it is a time consuming step.}

\item{ci_type}{Type of confidence interval to compute. The default value is "all". Other options are "set" for ci on each set, "t1" for ci on nominal catch by species, "t1-fmod" for ci on nominal catch by species and fishing mode "t2" and "t2-fmod" for ci by 1 degree square and month. A vector of several ci option can be provided. ci_type are computed only if  the ci parameter is TRUE.}

\item{Nboot}{Object of type \code{\link[base]{numeric}} expected. The number of bootstrap samples desired for the ci computation. The default value is 10.}

\item{plot_predict}{Object of type \code{\link[base]{logical}} expected. Logical indicating whether maps of catch at size have to be done.}
}
\description{
Run the t3 process, with the possibility to run all the process or stop.
}
\details{
For the argument "process", you can choose between 4 modalities (descending size classification):
\itemize{
 \item{all: } argument by default, you launch all the process
 \item{level1: } you launch data model initialisation and the process level 1
 \item{level2: } you launch data model initialisation and the process level 2
 \item{until_level2: } you launch data model initialisation, the process level 1 and 2
}
}
\seealso{
\code{\link[t3]{t3_level1}},  \code{\link[t3]{t3_level2}}, \code{\link[t3]{t3_level3}}, \code{\link[t3]{full_trips}}
}
