% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R, R/trip.R
\docType{data}
\name{trip}
\alias{trip}
\title{R6 class trip creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trip

trip
}
\description{
Create R6 reference object class trip

Create R6 reference object class trip
}
\keyword{datasets}
