% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_trips.R, R/trips.R
\docType{data}
\name{trips}
\alias{trips}
\title{R6 class trips creation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trips

trips
}
\description{
Create R6 reference object class trips

Create R6 reference object class trips
}
\keyword{datasets}
