% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well}
\alias{well}
\title{R6 class well creation}
\description{
Create R6 reference object class well
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{well$new()}}
\item \href{#method-clone}{\code{well$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 well class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{well$new(
  trip_id,
  well_id,
  well_minus10_weigth,
  well_plus10_weigth,
  well_global_weigth
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{(character) Trip identification.}

\item{\code{well_id}}{(character) Well identification.}

\item{\code{well_minus10_weigth}}{(integer) Catch weight of individus less than 10 tonnes (by well, in tonne, all species considerated).}

\item{\code{well_plus10_weigth}}{(integer) Catch weight of individus more than 10 tonnes (by well, in tonne, all species considerated).}

\item{\code{well_global_weigth}}{(integer) Catch weight of individus (less and more 10 tonnes categories, by well, in tonne, all species considerated).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{well$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
