% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well}
\alias{well}
\title{R6 class well}
\description{
Create R6 reference object class well
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-well-new}{\code{well$new()}}
\item \href{#method-well-clone}{\code{well$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-well-new"></a>}}
\if{latex}{\out{\hypertarget{method-well-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 well class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{well$new(
  trip_id,
  well_id,
  well_minus10_weight,
  well_plus10_weight,
  well_global_weight
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. Trip identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.}

\item{\code{well_minus10_weight}}{Object of class {\link[base]{integer}} expected. Catch weight of individuals less than 10 tonnes (by well, in tonne, considering all species).}

\item{\code{well_plus10_weight}}{Object of class {\link[base]{integer}} expected. Catch weight of individuals more than 10 tonnes (by well, in tonne, considering all species).}

\item{\code{well_global_weight}}{Object of class {\link[base]{integer}} expected. Catch weight of individuals (less and more 10 tonnes categories, by well, in tonne, considering all species).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-well-clone"></a>}}
\if{latex}{\out{\hypertarget{method-well-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{well$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
