% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellset.R
\name{wellset}
\alias{wellset}
\title{R6 class wellset}
\description{
Create R6 reference object class wellset
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-wellset-new}{\code{wellset$new()}}
\item \href{#method-wellset-clone}{\code{wellset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wellset-new"></a>}}
\if{latex}{\out{\hypertarget{method-wellset-new}{}}}
\subsection{Method \code{new()}}{
Initialize function for R6 wellset class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wellset$new(
  trip_id,
  activity_id,
  well_id,
  sample_id,
  weighted_weight,
  weighted_weight_minus10,
  weighted_weight_plus10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trip_id}}{Object of class {\link[base]{character}} expected. Trip identification.}

\item{\code{activity_id}}{Object of class {\link[base]{character}} expected. Activity identification.}

\item{\code{well_id}}{Object of class {\link[base]{character}} expected. Well identification.}

\item{\code{sample_id}}{Object of class {\link[base]{character}} expected. Sample identification.}

\item{\code{weighted_weight}}{Object of class {\link[base]{numeric}} expected. Set weight weighted by all set in the well(s).}

\item{\code{weighted_weight_minus10}}{Object of class {\link[base]{numeric}} expected. Set weight of individuals less than 10 kg (weighted by all set in the well(s))}

\item{\code{weighted_weight_plus10}}{Object of class {\link[base]{numeric}} expected. Set weight of individuals plus than 10 kg (weighted by all set in the well(s))}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wellset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-wellset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{wellset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
